% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{apply_digits}
\alias{apply_digits}
\title{Function to apply rounding on columns}
\usage{
apply_digits(x, digits, f = signif)
}
\arguments{
\item{x}{Data frame}

\item{digits}{Named vector with number of digits. Names correspond to the columns in \code{x} to round.}

\item{f}{Rounding function. Either named vector of rounding functions or function.}
}
\value{
Numeric vector with rounded values
}
\description{
Note: Need to check that attributes are not lost during rounding!!!
}
\examples{
\dontrun{
conc_rounded <- apply_digits(
  conc,
  digits = c(MEAN = 3, SD = 4, GMEAN = 3, GSD = 4)
)
}
}
