% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{AUC_IQnca}
\alias{AUC_IQnca}
\title{AUC calculation}
\usage{
AUC_IQnca(x, y, AUCMETHD = "Linear Log", TMAX = NULL, last = TRUE)
}
\arguments{
\item{x}{time or similar vector}

\item{y}{concentration or similar vector}

\item{AUCMETHD}{Defines AUC calculation method. "Linear Log", "LinearUp LogDown", "Linear LinearInterpolation", "Linear LinearLogInterpolation"}

\item{TMAX}{TMAX can be provided. If not provided Tmax will be assumed to be included in the provided data and can be determined. Otherwise
TMAX can be provided.}

\item{last}{if TRUE only return last value}
}
\value{
A list with entry for AUC and AUMC. Either a scalar (last) or a vector (all values)
}
\description{
Calculation of AUC by selected method.
Same methods as in Winnonlin are implemented: "Linear Log", "LinearUp LogDown", "Linear LinearInterpolation", "Linear LinearLogInterpolation"
The "Trapezoidal" dose not appear in the name of the method to make it shorter.
}
