% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_explode}
\alias{aux_explode}
\title{Splits a string based on a separator substring}
\usage{
aux_explode(input, separator = ",")
}
\arguments{
\item{input}{A string to split}

\item{separator}{A substring to use as separator}
}
\value{
A character vector with the split elements
}
\description{
By default a "," is used as separator.
}
