% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_explodePC}
\alias{aux_explodePC}
\title{Splits strings by separator only if separators not in parentheses}
\usage{
aux_explodePC(input, separator = ",", group = "round")
}
\arguments{
\item{input}{A string to split}

\item{separator}{A substring to use as separator}

\item{group}{"round", "square", or "curly" allowing grouping by parantheses
of defined type in which elements are not exploded}
}
\value{
A character vector with the split elements
}
\description{
By default a "," is used as separator.
}
