% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_fileparts}
\alias{aux_fileparts}
\title{aux_fileparts function}
\usage{
aux_fileparts(filename.with.path)
}
\arguments{
\item{filename.with.path}{path to file}
}
\description{
Writes a formatted string to a text file
}
\examples{
aux_fileparts("/path/to/file.csv")
}
