% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_fileread}
\alias{aux_fileread}
\title{aux_fileread function}
\usage{
aux_fileread(filename, collapserows = TRUE)
}
\arguments{
\item{filename}{filename possibly including path}

\item{collapserows}{indicates if rows should be collapsed to a string (default: TRUE)}
}
\description{
Reads contents of a text file
}
