% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_filewrite}
\alias{aux_filewrite}
\title{aux_filewrite function}
\usage{
aux_filewrite(text, filename)
}
\arguments{
\item{text}{text to write}

\item{filename}{filename possibly including path}
}
\description{
Writes a formatted string to a text file
}
