% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_mkdir}
\alias{aux_mkdir}
\title{aux_mkdir function}
\usage{
aux_mkdir(pathdir)
}
\arguments{
\item{pathdir}{path to folder to create}
}
\description{
Creates a folder if it does not yet exist
}
