% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_auxiliaries.R
\name{aux_plotlayout}
\alias{aux_plotlayout}
\title{Determine layout for distribution single plots to rows, columns, and pages}
\usage{
aux_plotlayout(nplots, nrow = NULL, ncol = NULL, npage = NULL)
}
\arguments{
\item{nplots}{Number of plots to distribute}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{npage}{Number of pages}
}
\value{
list with nrow, ncol, npage
}
\description{
If both nrow and ncol are given, npage is determined appropriately.
For other cases missing nrow/ncol is determined to match the required npage or to distribute on as many pages as possible.
\itemize{
\item either nrow or ncol is missing: assume 1 for the missing input, determine npage accordingly or use the given number of pages.
\item both nrow and ncol are missing: determined to match npage and keep an ratio of nrow/ncol approx 3/4 or general one page per plot
}
}
\author{
Anne Kümmel, IntiQuan
}
