% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_postFillChar}
\alias{aux_postFillChar}
\title{Postpad string with char to defined length}
\usage{
aux_postFillChar(value2postfill, lengthString, fillChar)
}
\arguments{
\item{value2postfill}{String to post fill}

\item{lengthString}{Length of final string}

\item{fillChar}{Char to fill with}
}
\description{
Postpad string with char to defined length
}
