% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_preFillChar}
\alias{aux_preFillChar}
\title{Prepad string with char to defined length}
\usage{
aux_preFillChar(value2prefill, lengthString, fillChar)
}
\arguments{
\item{value2prefill}{String to pre fill}

\item{lengthString}{Length of final string}

\item{fillChar}{Char to fill with}
}
\description{
Prepad string with char to defined length
}
