% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_simplifypath}
\alias{aux_simplifypath}
\title{Remove unnecessary dot-dot from (relative) paths.}
\usage{
aux_simplifypath(path)
}
\arguments{
\item{path}{Path character string}
}
\value{
string with simplified path (no .. within path)
}
\description{
Remove unnecessary dot-dot from (relative) paths.
}
\examples{
aux_simplifypath("this/does/not/../work")
}
