% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_splitVectorEqualPieces}
\alias{aux_splitVectorEqualPieces}
\title{Split vector into equal pieces}
\usage{
aux_splitVectorEqualPieces(x, n)
}
\arguments{
\item{x}{Vector to be splitted}

\item{n}{Number of elements per piece}
}
\value{
List with the split pieces
}
\description{
Splits a vector in pieces with defined length and
a remainder. Useful when to split datasets by ID for plotting over
several pages.
}
\examples{
x <- 1:20
aux_splitVectorEqualPieces(x,3)
}
