% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_strrep}
\alias{aux_strrep}
\title{Replaces some string in a string}
\usage{
aux_strrep(origstr, oldsubstr, newsubstr)
}
\arguments{
\item{origstr}{A string to replace something in}

\item{oldsubstr}{Substring to replace}

\item{newsubstr}{Substring to use for replacement}
}
\value{
Updated string
}
\description{
Legacy function ... not using regexpr
}
