% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_strtrim}
\alias{aux_strtrim}
\title{Trims a strings leading and trailing white spaces}
\usage{
aux_strtrim(input)
}
\arguments{
\item{input}{A string to trim}
}
\value{
Trimmed string
}
\description{
Trims a strings leading and trailing white spaces
}
