% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{aux_version}
\alias{aux_version}
\title{Check version function for packages}
\usage{
aux_version(
  pkgName = "IQnca",
  IQdesktop = NULL,
  exactVersion = NULL,
  minVersion = NULL,
  Rversion = NULL,
  isRopen = NULL,
  OS = NULL
)
}
\arguments{
\item{pkgName}{String with name of package (defaults to IQnca)}

\item{IQdesktop}{String defining the IQdesktop version used for the analysis}

\item{exactVersion}{String with exact version of package required}

\item{minVersion}{String with minimum version of package required}

\item{Rversion}{The user can provide the R version used for this analysis.
If provided (example: Rversion="3.5.1") then it will be checked if the current R
version matches it. If not, there will be an error message.}

\item{isRopen}{The user can define whether R Open is to be used for the analysis (TRUE or FALSE).
If provided then it will be checked if this is correct. If not, there will be an error message.}

\item{OS}{The user can provide which operating system used for this analysis.
If provided (example: OS="unix") then it will be checked if the OS matches it.
If not, there will be an error message.}
}
\value{
Version of package as string (if minVersion=NULL)
TRUE if version of package >= minVersion
Error message if version of package < minVersion
}
\description{
Check version function for packages
}
