% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{calc_stats_raw}
\alias{calc_stats_raw}
\title{Calculation of various statistics}
\usage{
calc_stats_raw(d, value_col)
}
\arguments{
\item{d}{data frame over which rows summary stats will be calculated}

\item{value_col}{column in data frame which is summarized}
}
\value{
data frame
}
\description{
Returns data frame with unformatted, numerical results
}
