% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{calc_summary_stats}
\alias{calc_summary_stats}
\title{Calculate statistics for a value column in a data frame}
\usage{
calc_summary_stats(d, value_col, digits = NULL, roundfun = signif)
}
\arguments{
\item{d}{data frame}

\item{value_col}{character indicating the column in the data frame to be summarized}

\item{digits}{Number of digits to to be used for rounding (if none given, the default of the applied function is used.).
Single numeric to be applied to all. Named numeric vector to specify digits for stats separately.
Names need to correspond to the keyword for the respective statistic.}

\item{roundfun}{Function or list of functions to be used for rounding (defaults to \link{signif}). If a list is given, names need to correspond
to the keyword for the respective statistic.
It is assumed that the given rounding function has two inputs: numeric to be rounded integer defining the number of digits used for rounding}
}
\value{
data frame with one row containing the statistics for value column of input dataset
}
\description{
The function calculates various statistics and rounds the values to the given number of digits.
Statistics that cannot be determined are given as "NC". The following statistics are calculated:
}
\details{
\tabular{ll}{
   Keyword \tab Description \cr
   MEAN \tab Arithmetic mean \cr
   SD \tab Standard deviation (arithmetic) \cr
   SE \tab Standard error (arithmetic) \cr
   CVPMN \tab Coefficient of variation in \% (arithmetic) \cr
   CI95MN \tab 95\%-confidence interval for arithmetic mean \cr
   CI90MN \tab 90\%-confidence interval for arithmetic mean \cr
   GMEAN \tab Geometric mean \cr
   GSD \tab Standard error (geometric) \cr
   CVPGM \tab Coefficient of variation in \% (geometric) \cr
   CI95GM \tab 95\%-confidence interval for geometric mean \cr
   CI90GM \tab 90\%-confidence interval for geometric mean \cr
   MEDIAN \tab Median \cr
   MIN \tab Minimum value \cr
   MAX \tab Maximum value \cr
}
}
\examples{
\dontrun{
par_stats <- calc_summary_stats(
  dataNCA,
  value_col = "CONC",
  digits = c(MEAN = 3, SD = 4, GMEAN= 3, GSD = 4))
}
}
