% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{check_IQdataNCA}
\alias{check_IQdataNCA}
\title{Check IQdataNCA}
\usage{
check_IQdataNCA(data)
}
\arguments{
\item{data}{IQdataNCA object}
}
\value{
Error if something is wrong
}
\description{
General checks for consistency
More will be added over time!
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()},
\code{\link{summaryObservations_IQdataNCA}()}
}
\concept{NCA Data}
