% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{compareIQRWNL_IQnca}
\alias{compareIQRWNL_IQnca}
\title{Simple pragmatic comparison function for IQnca and Winnonlin results}
\usage{
compareIQRWNL_IQnca(resIQR, WNLresFile, TOL = 1e-10, mapAUCINTnames = NULL)
}
\arguments{
\item{resIQR}{result from IQnca}

\item{WNLresFile}{Pivoted WinNonlin parameter file saved as CSV}

\item{TOL}{allowed relative difference}

\item{mapAUCINTnames}{named vector mapping IQR AUCINT column names to Winnonlin col names}
}
\description{
Simple pragmatic comparison function for IQnca and Winnonlin results
}
