% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_auxiliaries.R
\name{createPages_IQRoutputFigure}
\alias{createPages_IQRoutputFigure}
\title{Create pages from IQRoutputFigure}
\usage{
createPages_IQRoutputFigure(
  x,
  nrow = NULL,
  ncol = NULL,
  npage = NULL,
  legend.option = c("as.is", "remove", "common"),
  legend.object = NULL,
  legend.position = "right",
  legend.relsize = 0.2,
  title.relheight = 0.05,
  subtitle.relheight = 0.05,
  footer.relheight = 0.05
)
}
\arguments{
\item{x}{IQRoutputFigure object}

\item{nrow}{number of rows per page}

\item{ncol}{number of columns per page}

\item{npage}{number of pages. If nrow and ncol is given, this input is ignored.}

\item{legend.option}{Character or numeric whether to leave legend in plots as is ('as.is', 1),
remove all legends from the plots ('remove', 2), or plot common legend ('common', 3).
As common legend, the legend of the first plot per page is used if not user provided
by 'legend.object'}

\item{legend.object}{User-provided legend}

\item{legend.position}{Position of legend relative to main plots.
Can be 'right' (default), 'left', 'top', or 'bottom'.}

\item{legend.relsize}{Fraction of plot region (width for left or right position,
height for top or bottom position) allocated to legend.}

\item{title.relheight}{Fraction of page height allocated to plot title (if exists).}

\item{subtitle.relheight}{Fraction of page height allocated to plot subtitle (if exists).}

\item{footer.relheight}{Fraction of page height allocated to plot footer (if exists).}
}
\value{
list of plot objects or single plot object
}
\description{
Creates plot objects using \code{cowplot} arranging the plots from the
\link{IQRoutputFigure} into rows and columns and potentially multiple
pages. The title, subtitle, footer, and legend are plotted on each
created page.
}
\details{
Per default, plots are arranged to one page and a legend is plotted
on the right-hand side of the plots. The user can provide a plot object
of the legend. If not provided the legend is taken from the first plot
object in \code{content}.
If options for the layout (\code{opt.layout}) are given in the IQRoutputFigure
they will be applied unless a value is given as input.
Note that potential page dimention and resolution settings (\code{opt.page})
do not apply here.
}
\examples{
\dontrun{
 # Arrange plots in IQRouputFigure 'figobj' to 2 rows and 3 columns
 createPages_IQRoutputFigure(figobj, nrow = 2, ncol = 3)
 # Arrange plots in IQRouputFigure 'figobj' to 2 rows and as many
 # columns as needed to result in at most 3 pages
 createPages_IQRoutputFigure(figobj, nrow = 2, npage = 3)
}
}
\author{
Anne Kümmel, IntiQuan
}
