% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{dataSpec_IQdataNCA}
\alias{dataSpec_IQdataNCA}
\title{Show NCA data specification and display as a table}
\usage{
dataSpec_IQdataNCA()
}
\description{
Derived directly from the Excel data spec ... this is more flexible!
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()},
\code{\link{summaryObservations_IQdataNCA}()}
}
\concept{NCA Data}
