% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprogramming_tools.R
\name{date2datetime_IQRdataProgramming}
\alias{date2datetime_IQRdataProgramming}
\title{Conversion of date/time in string format to time in string format}
\usage{
date2datetime_IQRdataProgramming(
  dateString,
  format = "\%Y-\%m-\%dT\%H:\%M",
  formattime = "\%H:\%M"
)
}
\arguments{
\item{dateString}{Vector of dates/times in string format}

\item{format}{Format of the date/time information, defined as the format for as.POSIXxt()}

\item{formattime}{Format for the time string, defined as format for as.POSIXct()}
}
\value{
Vector of time strings
}
\description{
Conversion of date/time in string format to time in string format
}
