% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprogramming_tools.R
\name{date2time_IQRdataProgramming}
\alias{date2time_IQRdataProgramming}
\title{Conversion of date/time in string format to numeric values}
\usage{
date2time_IQRdataProgramming(dateString, format = "\%Y-\%m-\%dT\%H:\%M")
}
\arguments{
\item{dateString}{Vector of dates/times in string format}

\item{format}{Format of the date/time information, defined as the format for as.POSIXxt()}
}
\value{
List with the elements time_seconds, time_minutes, time_hours, time_days, time_weeks, time_years
}
\description{
Conversion of date/time in string format to numeric values
}
