% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{doc_IQnca}
\alias{doc_IQnca}
\title{Open IQnca documentation in browser}
\usage{
doc_IQnca()
}
\description{
Open IQnca documentation in browser
}
\seealso{
Other Help & Documentation: 
\code{\link{examples_IQnca}()},
\code{\link{install_MIDDmodule}()},
\code{\link{install_workshop}()},
\code{\link{support}()}
}
\concept{Help & Documentation}
