% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{examples_IQnca}
\alias{examples_IQnca}
\title{Install examples to allow running examples from the book}
\usage{
examples_IQnca(path = "~/IntiQuan/IQnca")
}
\arguments{
\item{path}{Path to where to copy the book example files. Should not contain spaces}
}
\description{
This function copies the examples in inst folder
to the default path (~/IntiQuan/IQnca/examples). In addition it sets the
path to the path where the material was copied.
}
\seealso{
Other Help & Documentation: 
\code{\link{doc_IQnca}()},
\code{\link{install_MIDDmodule}()},
\code{\link{install_workshop}()},
\code{\link{support}()}
}
\concept{Help & Documentation}
