% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{export_IQdataNCA}
\alias{export_IQdataNCA}
\title{Export an IQdataNCA object}
\usage{
export_IQdataNCA(data, filename = "nca_data")
}
\arguments{
\item{data}{IQdataNCA object}

\item{filename}{Filename with path for export}
}
\description{
Essentially the dataset is exported as CSV with an attributes file.
All information is self contained in the CSV file ... so the attributes file at the moment
does not play a real role and is not produced upon export. .
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()},
\code{\link{summaryObservations_IQdataNCA}()}
}
\concept{NCA Data}
