% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{export_IQRrmd}
\alias{export_IQRrmd}
\title{Export IQRrmd object to an rmd file}
\usage{
export_IQRrmd(rmdDoc, filename)
}
\arguments{
\item{rmdDoc}{IQRrmd object}

\item{filename}{Filename (might include path). If not already appended, ".rmd"
will be appended.}
}
\description{
Export IQRrmd object to an rmd file
}
