% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{exportCDISC_IQnca}
\alias{exportCDISC_IQnca}
\title{Export an IQnca object to CDISC PP and ADPP}
\usage{
exportCDISC_IQnca(
  data,
  pathname = ".",
  SIGNIF = 8,
  addColLabels = NULL,
  parameterReport = "standard"
)
}
\arguments{
\item{data}{IQnca object}

\item{pathname}{Path to which the export happens. Default filenames pp.xpt and adpp.xpt will be used. Path can be defined.}

\item{SIGNIF}{Number of significant digits}

\item{addColLabels}{Named list. Names define column names and elements define labels for these columns - only used in generation of ADNCA}

\item{parameterReport}{Character string defining the PK parameters to report in the listing.
By default these parameters will be selected based on the "Standard" parameters defined in the IQparamNCA_Specification.xlsx
document, subject to the profile and administration type. The user can select "All" or "Standard".
Alternatively, parameterReport can also be a vector or parameter names (based on PKPARAMCD in IQparamNCA_Specification.xlsx).
Instead of the standard or all ones then these will be considered in the listing (also subject to profile and
administration type). Note that interval AUC (if calculated is always added by default!)
If defined it overrides the default labels.}
}
\description{
Export the NCA results to PP and ADPP XPT files. Still somewhat exploratory but reasonable.
Default filenames pp.xpt and adpp.xpt will be used. Path can be defined.
}
\seealso{
Other NCA Parameter Export: 
\code{\link{export_IQnca}()}
}
\concept{NCA Parameter Export}
