% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_classes_IQdataNCA.R
\name{figure_footer}
\alias{figure_footer}
\title{Internal function defining BLQ footer text for figures.}
\usage{
figure_footer(x)
}
\arguments{
\item{x}{string defining setting for BLQ handling}
}
\value{
Footer text
}
\description{
Internal function defining BLQ footer text for figures.
}
