% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_classes_IQdataNCA.R
\name{figure_lines_IQdataNCA}
\alias{figure_lines_IQdataNCA}
\title{Generate line figures of PK concentration for reporting}
\usage{
figure_lines_IQdataNCA(
  data,
  split_by = "PROFILE",
  lines_by = "GROUP",
  panel_by = "STUDYID",
  dosenorm = FALSE,
  time = "asis",
  use_TAFD = TRUE,
  figure_number = NULL,
  figure_head = NULL,
  labels = NULL,
  yscale = c("linlog", "lin", "log"),
  legend_pos = "bottom",
  sameXaxis = TRUE,
  sameYaxis = TRUE,
  filename = NULL,
  type = "overlay_figure",
  ...
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{split_by}{Column to split figures to separate pages.}

\item{lines_by}{grouping used for linetypes}

\item{panel_by}{variable to create panels by}

\item{dosenorm}{logical whether to plot dose normalized concentrations}

\item{time}{Character string to defined the time information used ("asis", "nominal", "actual").
"asis" will use the selected time information in the dataset (TIME or TAFD). "nominal" will use NTIME or NTAFD.
"actual" will use ATIME or ATAFD.}

\item{use_TAFD}{logical. FALSE (default) uses TIME, NTIME, or ATIME, depending on 'time' settings. TRUE will use
TAFD, NTAFD, or ATAFD.}

\item{figure_number}{Character string with figure number information, added to the title for each figure, if defined}

\item{figure_head}{Column names of covariates unique within pages to be displayed above the table}

\item{labels}{names character vector with labels for columns}

\item{yscale}{character identifying whether to plot linear ("lin")
or semilogarithmic ("log") or both views ("linlog")}

\item{legend_pos}{specifying legend position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{sameXaxis}{Flag whether to use same x axis scale across pages (splits). Defaults to TRUE.}

\item{sameYaxis}{Flag whether to use same y axis scale across pages (splits). Defaults to TRUE.}

\item{filename}{Name of PDF file to generate with plots}

\item{type}{Type of figure (meta data)}

\item{...}{Arguments to \link{IQRoutputFigure} to adjust plotting settings}
}
\value{
Returns a list of all plots if filename is NULL
}
\description{
Lines connect data for each profile within one individual. Lines can be color-stratified and distributed to different panels.
}
\details{
Ignored records (IGNORER) are removed from these plots. IGNORSUM and IGNORNCA records are included.
Ignored subjects (IGNOREI) are included. In addition to a PDF file with the figures a folder containing
amongst other meta information an rmd file will be generated. The RMD file is for seamless reporting in Word with IQReport.
}
\seealso{
Other TLF customized: 
\code{\link{figure_indivConc_IQdataNCA}()},
\code{\link{figure_summary_IQdataNCA}()},
\code{\link{listing_PKconc_IQdataNCA}()},
\code{\link{listing_PKpars_IQnca}()},
\code{\link{table_summary_PKconc_IQdataNCA}()},
\code{\link{table_summary_PKpars_IQnca}()}
}
\concept{TLF customized}
