% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{figure_spaghetti_IQdataNCA}
\alias{figure_spaghetti_IQdataNCA}
\title{Plot summary spaghetti plots of IQdataNCA objects (non-dose-normalized)}
\usage{
figure_spaghetti_IQdataNCA(
  data,
  figurenumber = NULL,
  logY = TRUE,
  strat = "USUBJID",
  scaleXsame = TRUE,
  scaleYsame = TRUE,
  filename = "spaghetti.pdf",
  nindiv = 1,
  FLAGTIME = "nominal",
  USETAD = TRUE,
  alpha = 0.8
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{figurenumber}{Character string with figure number information, added to the title for each figure, if defined}

\item{logY}{Log y axis is TRUE, otherwise linear Y axis}

\item{strat}{Character. Name of column to use for stratification (default: NULL => no stratification). The only requirement
is that the entries in the stratification column are unique per subject per profile. The columns can contain
numeric of aphanumeric entries. They can be categorical or continuous. However, it is advisable for the purpose
of useful plots to use continuous numeric columns only for stratification when there is a certain limited
number of unique values. But it is your call entirely :-)}

\item{scaleXsame}{logical. It TRUE then the same min anx max value on the X-axis will be used for each plot}

\item{scaleYsame}{logical. It TRUE then the same min anx max value on the Y-axis will be used for each plot}

\item{filename}{Name of PDF file to generate with plots}

\item{nindiv}{Number of plots per page if exported to PDF}

\item{FLAGTIME}{Character string to defined the time information used ("asis", "nominal", "actual").
"asis" will use the selected time information in the dataset (TIME or TAFD). "nominal" will use NTIME or NTAFD.
"actual" will use ATIME or ATAFD.}

\item{USETAD}{logical. TRUE uses TIME, NTIME, or ATIME, depending on FLAGTIME settings. FALSE will use
TAFD, NTAFD, or ATAFD.}

\item{alpha}{alpha value for the plotting.}
}
\value{
Returns a list of all plots if filename is NULL
}
\description{
Stratified by GROUP and PROFILE in the data. Ordered by GROUP(N) and then PROFILE. Default colored by USUBJID.
Only ignored records (IQGNORER) are excluded from these plots. Stratification in terms of coloring by selected column
can be done.
}
\seealso{
Other NCA Data Figures: 
\code{\link{figure_spaghetti_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_geomean_IQdataNCA}()},
\code{\link{figure_summary_geommean_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_mean_IQdataNCA}()},
\code{\link{figure_summary_mean_dosenorm_IQdataNCA}()},
\code{\link{figures_indiv_IQdataNCA}()},
\code{\link{plot_summary_IQdataNCAsum}()}
}
\concept{NCA Data Figures}
