% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_classes_IQdataNCA.R
\name{figure_summary_IQdataNCA}
\alias{figure_summary_IQdataNCA}
\title{Generate summary figures of PK concentration for reporting}
\usage{
figure_summary_IQdataNCA(
  data,
  statistic_avg = c("MEAN", "GMEAN", "MEDIAN"),
  statistic_var = c("SD", "GSD", "PI90", "PI95"),
  split_by = "PROFILE",
  lines_by = "GROUP",
  panel_by = "STUDYID",
  dosenorm = FALSE,
  use_TAFD = FALSE,
  figure_number = NULL,
  figure_head = NULL,
  labels = NULL,
  yscale = "linlog",
  legend_pos = "bottom",
  sameXaxis = TRUE,
  sameYaxis = TRUE,
  filename = NULL,
  type = "figure_summary",
  ...
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{statistic_avg}{Statistic to be used to present 'average' value. MEAN, GMEAN, or MEDIAN. See description.}

\item{statistic_var}{Statistic to be used to present 'variablity' of values. SD, GSD, PI90, or PI95. See description.}

\item{split_by}{Column to split figures to separate pages.}

\item{lines_by}{grouping used for linetypes}

\item{panel_by}{variable to create panels by}

\item{dosenorm}{logical whether to plot dose normalized concentrations}

\item{use_TAFD}{logical. FALSE (default) uses TIME, NTIME, or ATIME, depending on 'time' settings. TRUE will use
TAFD, NTAFD, or ATAFD.}

\item{figure_number}{Character string with figure number information, added to the title for each figure, if defined}

\item{figure_head}{Column names of covariates unique within pages to be displayed above the table}

\item{labels}{names character vector with labels for columns}

\item{yscale}{character identifying whether to plot linear ("lin")
or semilogarithmic ("log") or both views ("linlog")}

\item{legend_pos}{specifying legend position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{sameXaxis}{Flag whether to use same x axis scale across pages (splits). Defaults to TRUE.}

\item{sameYaxis}{Flag whether to use same y axis scale across pages (splits). Defaults to TRUE.}

\item{filename}{Name of PDF file to generate with plots}

\item{type}{Type of figure (meta data)}

\item{...}{Arguments to \link{IQRoutputFigure} to adjust plotting settings}
}
\value{
Returns a list of all plots if filename is NULL
}
\description{
An average and a variability statistic are calculated for the stratification defined by the
split to pages (\code{split_by}), linetype stratification (\code{lines_by}), and facetting to different panels (\code{panel_by}).
The average and variability can be defined by the user:
}
\details{
\tabular{lll}{
   Input \tab Keyword \tab Description \cr
   statistic_avg \tab MEAN \tab Arithmetic mean \cr
   statistic_avg \tab GMEAN \tab Geometric mean \cr
   statistic_avg \tab MEDIAN \tab Median \cr
   statistic_var \tab SD \tab Standard deviation (arithmetic) \cr
   statistic_var \tab GSD \tab Standard deviation (geometric) \cr
   statistic_var \tab PI90 \tab 90\%-interval \cr
   statistic_var \tab PI95 \tab 95\%-interval \cr
}


Ignored records (IGNORER) are removed from these plots. IGNORSUM and IGNORNCA records are included.
Ignored subjects (IGNOREI) are included. In addition to a PDF file with the figures a folder containing
amongst other meta information an rmd file will be generated. The RMD file is for seamless reporting in Word with IQReport.
}
\seealso{
Other TLF customized: 
\code{\link{figure_indivConc_IQdataNCA}()},
\code{\link{figure_lines_IQdataNCA}()},
\code{\link{listing_PKconc_IQdataNCA}()},
\code{\link{listing_PKpars_IQnca}()},
\code{\link{table_summary_PKconc_IQdataNCA}()},
\code{\link{table_summary_PKpars_IQnca}()}
}
\concept{TLF customized}
