% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{figure_summary_mean_dosenorm_IQdataNCA}
\alias{figure_summary_mean_dosenorm_IQdataNCA}
\title{Plot of summary dose-normalized PK concentrations (arithmetic mean and SD)}
\usage{
figure_summary_mean_dosenorm_IQdataNCA(
  data,
  figurenumber = NULL,
  logY = TRUE,
  strat = "GROUP",
  scaleXsame = TRUE,
  scaleYsame = TRUE,
  filename = "summarymeandn.pdf",
  nindiv = 1,
  USETAD = TRUE,
  alpha = 0.8
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{figurenumber}{Character string with figure number information, added to the title for each figure, if defined}

\item{logY}{Log y axis is TRUE, otherwise linear Y axis}

\item{strat}{Character. Name of column to use for stratification (default: GROUP). The only requirement
is that the entries in the stratification column are unique per subject per profile. The columns can contain
numeric of aphanumeric entries. They can be categorical or continuous. However, it is advisable for the purpose
of useful plots to use continuous numeric columns only for stratification when there is a certain limited
number of unique values. But it is your call entirely :-)}

\item{scaleXsame}{logical. It TRUE then the same min anx max value on the X-axis will be used for each plot}

\item{scaleYsame}{logical. It TRUE then the same min anx max value on the Y-axis will be used for each plot}

\item{filename}{Name of PDF file to generate with plots}

\item{nindiv}{Number of plots per page if exported to PDF}

\item{USETAD}{logical. TRUE uses TIME, NTIME, or ATIME, depending on FLAGTIME settings. FALSE will use
TAFD, NTAFD, or ATAFD.}

\item{alpha}{alpha value for the plotting.}
}
\value{
Returns a list of all plots if filename is NULL
}
\description{
Always only for nominal time (NTIME or NTAFD)!
Since dose-normalization is done, standard stratifictation is done by PROFILE.
Additional stratification can be provided by the user and is shown in different colors.
Default stratification is by GROUP (ordered by GROUPN).
Ignored subjects (IGNOREI) and ignored records (IGNORER and IGNORSUM) are
excluded from these plots.
}
\seealso{
Other NCA Data Figures: 
\code{\link{figure_spaghetti_IQdataNCA}()},
\code{\link{figure_spaghetti_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_geomean_IQdataNCA}()},
\code{\link{figure_summary_geommean_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_mean_IQdataNCA}()},
\code{\link{figures_indiv_IQdataNCA}()},
\code{\link{plot_summary_IQdataNCAsum}()}
}
\concept{NCA Data Figures}
