% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{figure_summary_mean_IQdataNCA}
\alias{figure_summary_mean_IQdataNCA}
\title{Plot of summary PK concentrations (arithmetic mean and SD)}
\usage{
figure_summary_mean_IQdataNCA(
  data,
  figurenumber = NULL,
  logY = TRUE,
  scaleXsame = TRUE,
  scaleYsame = TRUE,
  filename = "summarymean.pdf",
  nindiv = 1,
  USETAD = TRUE,
  alpha = 0.8
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{figurenumber}{Character string with figure number information, added to the title for each figure, if defined}

\item{logY}{Log y axis is TRUE, otherwise linear Y axis}

\item{scaleXsame}{logical. It TRUE then the same min anx max value on the X-axis will be used for each plot}

\item{scaleYsame}{logical. It TRUE then the same min anx max value on the Y-axis will be used for each plot}

\item{filename}{Name of PDF file to generate with plots}

\item{nindiv}{Number of plots per page if exported to PDF}

\item{USETAD}{logical. TRUE uses NTIME. FALSE will use NTAFD}

\item{alpha}{alpha value for the plotting.}
}
\value{
Returns a list of all plots if filename is NULL
}
\description{
Always only for nominal time (NTIME)!
Stratified by PROFILE and GROUP (ordered by GROUPN). One plot per stratum.
Ignored subjects (IGNOREI) and ignored records (IGNORER and IGNORSUM) are
excluded from these plots.
}
\seealso{
Other NCA Data Figures: 
\code{\link{figure_spaghetti_IQdataNCA}()},
\code{\link{figure_spaghetti_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_geomean_IQdataNCA}()},
\code{\link{figure_summary_geommean_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_mean_dosenorm_IQdataNCA}()},
\code{\link{figures_indiv_IQdataNCA}()}
}
\concept{NCA Data Figures}
