% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{figures_indiv_IQdataNCA}
\alias{figures_indiv_IQdataNCA}
\title{Generate individual PK concentration figures for reporting}
\usage{
figures_indiv_IQdataNCA(
  data,
  figurenumber = NULL,
  logY = TRUE,
  filename = "figures_individual",
  nindiv = 1,
  FLAGTIME = "asis",
  USETAD = TRUE
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{figurenumber}{Character string with figure number information, added to the title for each figure, if defined}

\item{logY}{Log y axis is TRUE, otherwise linear Y axis}

\item{filename}{Name of PDF file to generate with plots}

\item{nindiv}{Number of plots per page if exported to PDF}

\item{FLAGTIME}{Character string to defined the time information used ("asis", "nominal", "actual").
"asis" will use the selected time information in the dataset (TIME or TAFD). "nominal" will use NTIME or NTAFD.
"actual" will use ATIME or ATAFD.}

\item{USETAD}{logical. TRUE uses TIME, NTIME, or ATIME, depending on FLAGTIME settings. FALSE will use
TAFD, NTAFD, or ATAFD.}
}
\value{
Returns a list of all plots if filename is NULL
}
\description{
Ignored records (IGNORER) are removed from these plots. IGNORSUM and IGNORNCA records are included.
Ignored subjects (IGNOREI) are included. In addition to a PDF file with the figures an rmd file will
be generated in the same folder and same filename (.rmd) extension). the RMD file is for seamless
reporting in Word with IQReport.
}
\seealso{
Other NCA Data Figures: 
\code{\link{figure_spaghetti_IQdataNCA}()},
\code{\link{figure_spaghetti_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_geomean_IQdataNCA}()},
\code{\link{figure_summary_geommean_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_mean_IQdataNCA}()},
\code{\link{figure_summary_mean_dosenorm_IQdataNCA}()}
}
\concept{NCA Data Figures}
