% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{geocv}
\alias{geocv}
\title{geometric CV auxiliary}
\usage{
geocv(x, na.rm = FALSE)
}
\arguments{
\item{x}{x}

\item{na.rm}{logical scalar indicating whether to remove missing values from x.
If na.rm=FALSE (the default) and x contains missing values, then a missing
value (NA) is returned. If na.rm=TRUE, missing values are removed from x
prior to computing the coefficient of variation.}
}
\value{
geometric CV
}
\description{
geometric CV auxiliary
}
