% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_auxiliaries.R
\name{geom_uperrorbar}
\alias{geom_uperrorbar}
\title{geom_uperrorbar ... nice to have}
\usage{
geom_uperrorbar(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{See geom_errorbar}

\item{data}{See geom_errorbar}

\item{stat}{See geom_errorbar}

\item{position}{See geom_errorbar}

\item{...}{See geom_errorbar}

\item{na.rm}{See geom_errorbar}

\item{show.legend}{See geom_errorbar}

\item{inherit.aes}{See geom_errorbar}
}
\description{
geom_uperrorbar ... nice to have
}
