% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{geomean}
\alias{geomean}
\title{geometric mean auxiliary}
\usage{
geomean(x, na.rm = FALSE)
}
\arguments{
\item{x}{x}

\item{na.rm}{FALSE or TRUE}
}
\value{
exp(mean(log(x))). Note that only positive values are considered!!!
}
\description{
geometric mean auxiliary
}
