% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_auxiliaries.R
\docType{data}
\name{GeomUperrorbar}
\alias{GeomUperrorbar}
\title{Needed for geom_uperrorbar}
\format{
An object of class \code{GeomUperrorbar} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomUperrorbar
}
\description{
Needed for geom_uperrorbar
}
\keyword{datasets}
