% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{geosd}
\alias{geosd}
\title{geometric sd auxiliary}
\usage{
geosd(x, na.rm = FALSE, sqrt.unbiased = TRUE)
}
\arguments{
\item{x}{x}

\item{na.rm}{logical scalar indicating whether to remove missing values from x.
If na.rm=FALSE (the default) and x contains missing values, then a missing
value (NA) is returned. If na.rm=TRUE, missing values are removed from x
prior to computing the coefficient of variation.}

\item{sqrt.unbiased}{logical scalar specifying what method to use to compute
the sample standard deviation of the log-transformed observations. If
sqrt.unbiased=TRUE (the default), the square root of the unbiased estimator
of variance is used, otherwise the method of moments estimator of standard
deviation is used. See the DETAILS section for more information.}
}
\value{
geometric SD
}
\description{
geometric sd auxiliary
}
