% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{get_default_labels}
\alias{get_default_labels}
\title{Define default labels for columns for IQdataNCA object}
\usage{
get_default_labels(d)
}
\arguments{
\item{d}{IQdataNCA object}
}
\value{
Named vector of labels. Names are column names for which labels apply
}
\description{
Define default labels for columns for IQdataNCA object
}
\examples{
\dontrun{
column_labels <- get_default_labels(dataNCA)
}
}
