% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{getaddcovcolumns_IQdataNCA}
\alias{getaddcovcolumns_IQdataNCA}
\title{Get additional (undefined potential covariate) columns for IQdataNCA objects}
\usage{
getaddcovcolumns_IQdataNCA(data)
}
\arguments{
\item{data}{IQdataNCA object}
}
\description{
Get additional (undefined potential covariate) columns for IQdataNCA objects
}
