% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{getColumnNames_IQdataNCA}
\alias{getColumnNames_IQdataNCA}
\title{Get information on the type of columns and their names for an IQdataNCA object}
\usage{
getColumnNames_IQdataNCA()
}
\value{
List with named entries. Each entry a vector with column names belonging to the "name" type of columns
}
\description{
Derived directly from the Excel data spec ... this is more flexible!
}
