% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprogramming_tools.R
\name{getLabels_dataframe}
\alias{getLabels_dataframe}
\title{Extract label information of columns in data.frames}
\usage{
getLabels_dataframe(data, orderAlphabetically = FALSE, table = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{orderAlphabetically}{If TRUE then output ordered alphabetically by column names}

\item{table}{If TRUE then an IQRoutputTable is returned. Otherwise a data.frame}
}
\description{
Useful if data has been imported e.g. using IQRloadSASdata.
By default an IQRoutputTable object is returned for easier reading or reporting.
Alternatively a data.frame can be returned. Sorting is possible if desired.
}
