% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{getTIMEUNITname_IQdataNCA}
\alias{getTIMEUNITname_IQdataNCA}
\title{Get name of time unit in IQdataNCA object}
\usage{
getTIMEUNITname_IQdataNCA(data)
}
\arguments{
\item{data}{IQdataNCA object}
}
\value{
Returns the name of the unit ready for display in plots and tables and listings.
}
\description{
This is useful for display and annotation purposes in plots and tables.
Example "min", "hour", etc.
}
