% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{getTIMEUNITsymbol_IQdataNCA}
\alias{getTIMEUNITsymbol_IQdataNCA}
\title{Get symbol of time unit in IQdataNCA object}
\usage{
getTIMEUNITsymbol_IQdataNCA(data)
}
\arguments{
\item{data}{IQdataNCA object}
}
\value{
Returns the symbol of the unit ready for display in plots and tables and listings.
}
\description{
This is useful for use of the time unit in units and constructed units for
derived NCA PK parameters. Example "min", "h", etc.
}
