% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{handleTIMEplot_IQdataNCA}
\alias{handleTIMEplot_IQdataNCA}
\title{Function to define the time column and their labeling used for plotting}
\usage{
handleTIMEplot_IQdataNCA(
  d,
  FLAGTIME = "nominal",
  USETAD = TRUE,
  add_unit = TRUE
)
}
\arguments{
\item{d}{IQdataNCA object}

\item{FLAGTIME}{character being either "nominal" (nominal time), "actual" (actual time), or "asis" (use setting as pre-defined in dataset).}

\item{USETAD}{FLAG whether to use time after last dose (TRUE, default) or time after first dose (FALSE)}

\item{add_unit}{Flag whether to add unit to axis label (defaults to true)}
}
\value{
list with updated IQdataNCA object and time axis label
}
\description{
Column TIMEPLOT is added to the dataset which should be used when plotting the data.
The label for the time axis is defined.
}
