% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tools_IQReport.R
\name{has_IQReport_executable}
\alias{has_IQReport_executable}
\title{Check if IQReport is installed and available}
\usage{
has_IQReport_executable()
}
\value{
Returns TRUE if IQReport is available on the system
}
\description{
Check if IQReport is installed and available
}
