% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{ignoreNCA_IQdataNCA}
\alias{ignoreNCA_IQdataNCA}
\title{Define records to be excluded from NCA PK parameter determination}
\usage{
ignoreNCA_IQdataNCA(data, REASON, USUBJID, PROFILE = NULL, IX)
}
\arguments{
\item{data}{IQdataNCA object in which to ignore selected records}

\item{REASON}{Reason for ignoring the record(s)}

\item{USUBJID}{USUBJID of the subject from which records are to be ignored in the summary procedures. Can be character vector with multiple USUBJIDs}

\item{PROFILE}{Profile name (entry in PROFILE column) to which to apply the ignoring (if multiple profiles available in the IQdataNCA object).
If no PROFILE name defined then this is applied to the subject for all possible profiles. PROFILE can be character vector, allowing to define
multiple profiles to which the ignoring should be applied. If only one profile present then PROFILE is not required to be defined.}

\item{IX}{Index (IX column value) of the record(s) to be ignored in NCA parameter determination with the same reason. Can be scalar or vector.}
}
\value{
Updated IQdataNCA object with adjusted ignore settings
}
\description{
Ignoring of records is done by entering a reason in the IGNORNCA column. A record that is ignored
for the NCA parameter determination is not considered for slope calculation or any other PK parameter
determination. It will still be listed in the individual concentration listings and might appear in the
concentration summaries. After ignoring the slope is recalculated.
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()}
}
\concept{NCA Data}
