% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{ignoresubject_IQdataNCA}
\alias{ignoresubject_IQdataNCA}
\title{Define subjects to be ignored for analyses with reason why}
\usage{
ignoresubject_IQdataNCA(data, REASON, USUBJID, PROFILE = NULL)
}
\arguments{
\item{data}{IQdataNCA object in which to ignore selected subjects}

\item{REASON}{Reason for ignoring the subject(s)}

\item{USUBJID}{USUBJID of the subject to be ignored. Can be character vector with multiple USUBJIDs (same reason then)}

\item{PROFILE}{Profile name (entry in PROFILE column) to which to apply the ignoring (if multiple profiles available in the IQdataNCA object).
If no PROFILE name defined then this is applied to the subject for all possible profiles. PROFILE can be character vector, allowing to define
multiple profiles to which the ignoring should be applied.}
}
\value{
Updated IQdataNCA object with adjusted ignore settings
}
\description{
Ignoring of subjects is done by entering a reason in the IGNOREI column. A subject that is ignored
is not considered at all in any step but the listing of individual PK concentrations. This means it does not appear
in the PK parameter listings, tables, figures, and neither is considered in the PK parameter determination.
An ignored subject will appear in the listings of individual PK concentrations but it will be documented
in the text that it is ignored from the rest of the analysis.
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()}
}
\concept{NCA Data}
