% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{install_MIDDmodule}
\alias{install_MIDDmodule}
\title{Installs defined webinar module from the web}
\usage{
install_MIDDmodule(module_number)
}
\arguments{
\item{module_number}{Module number}
}
\description{
Convenient function to install IntiQuan Webinar material directly from the IntiQuan Server
}
\seealso{
Other Help & Documentation: 
\code{\link{doc_IQnca}()},
\code{\link{examples_IQnca}()},
\code{\link{install_workshop}()},
\code{\link{support}()}
}
\concept{Help & Documentation}
