% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{install_workshop}
\alias{install_workshop}
\title{Installs workshop material}
\usage{
install_workshop(workshop)
}
\arguments{
\item{workshop}{Workshop identifier}
}
\description{
Convenient function to install workshop material directly from the IntiQuan Server
}
\seealso{
Other Help & Documentation: 
\code{\link{doc_IQnca}()},
\code{\link{examples_IQnca}()},
\code{\link{install_MIDDmodule}()},
\code{\link{support}()}
}
\concept{Help & Documentation}
