% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{intervalAUC_NCA}
\alias{intervalAUC_NCA}
\title{Define interval AUC calculation}
\usage{
intervalAUC_NCA(colname, label, tstart, tend)
}
\arguments{
\item{colname}{Column name in the result}

\item{label}{Label for the exported output}

\item{tstart}{Start of the interval}

\item{tend}{End of the interval}
}
\value{
Just a list - can be used as input argument 'intervalAUC' to nca_IQdataNCA()
}
\description{
Helper function to construct the intervalAUC input argument for nca_IQdataNCA()
}
