% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{IQdataNCA}
\alias{IQdataNCA}
\title{Create an IQdataNCA object}
\usage{
IQdataNCA(
  data,
  COMPTYPE = "exogenous",
  FLAGTIME = "actual",
  FATIMIMP = "asis",
  FLGBLQPR = "0",
  FLGBLQIN = "missing",
  FLGBLQP1 = "LLOQ/2",
  FLGBLQPO = "missing",
  FGBQPLIN = "asCONC",
  FGBQPLOG = "asCONC",
  AUCMETHD = "Linear Log",
  AUCINVAL = NA,
  SLOPETOL = 1e-04,
  FLAGoverwrite = TRUE,
  FLAGignore = TRUE
)
}
\arguments{
\item{data}{data.frame with initial definitions}

\item{COMPTYPE}{Type of compound: "exogenous" or "endogenous"}

\item{FLAGTIME}{“actual”: actual time used in NCA PK parameter determination, “nominal”: nominal time used in NCA PK parameter determination.
Default: "actual" It will be checked if ATIME is defined. If not, then "nominal" will be used.}

\item{FATIMIMP}{Used in case if basis for analysis is actual time but some actual time point is missing.
In this case the user can decide to impute this time point if the nominal time is available.
“asis”  or “nominal”. “asis”: if missing then it will not be imputed and not be considered in the analysis.
If “nominal” then it will be imputed based on NTIME, NTAFD, NDUR and impact TIME, TAFD, and DUR. Default: "asis"}

\item{FLGBLQPR}{Handling BLLOQ values before first observation above LLOQ. Options: "asis", "0", " LLOQ/2", "LLOQ", "missing".}

\item{FLGBLQIN}{Handling BLLOQ values between observations above LLOQ. Options: "asis", "0", " LLOQ/2", " LLOQ ", "missing".}

\item{FLGBLQP1}{Handling first BLLOQ value after last observation above LLOQ. Options: "asis", "0", " LLOQ/2", " LLOQ ", "missing".}

\item{FLGBLQPO}{Handling BLLOQ values after first BLLOQ post last observation above LLOQ. Options: "asis", "0", " LLOQ/2", " LLOQ ", "missing".}

\item{FGBQPLIN}{Handling BLLOQ values in plots on linear Y axis "asCONC", asis", "0", " LLOQ/2", " LLOQ ", "missing".}

\item{FGBQPLOG}{Handling BLLOQ values in plots on log Y axis "asCONC", "asis", "0", " LLOQ/2", " LLOQ ", "missing".}

\item{AUCMETHD}{Defines AUC calculation method. "Linear Log", "LinearUp LogDown", "Linear LinearInterpolation", "Linear LinearLogInterpolation"}

\item{AUCINVAL}{Character string defining one or more AUC intervals to be calculated. Example: AUCINVAL <- "[0;24];[0;48]". Note that
separators are ";" to allow to store that information also in a CSV file!}

\item{SLOPETOL}{Tolerance for bestslope algorithm}

\item{FLAGoverwrite}{If TRUE then derived columns will be regenerated even if already provided in data (safest).
If FALSE then the information in the provided columns will be kept as-is (requires to know exectly what to do).}

\item{FLAGignore}{If TRUE then default rules for exclusion of records from summary tables and NCA parameter calculation
are used when applicable (e.g. pre-dose samples >LLOQ). If FALSE then all samples by default are passed through as
available in the input data. The first setting is for real use. The second is mainly to allow comparison with results from other
NCA tools.}
}
\value{
IQdataNCA object
}
\description{
Generates an IQdataNCA object. The minimum required columns (see getColumnNames_IQdataNCA) need to be present.
Initial slope calculation will be conducted as well, based on the bestslope algorithm. This can later be refined
but is useful to get a quick understanding of the data.
}
\seealso{
Other NCA Data: 
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()}
}
\concept{NCA Data}
