% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings_plots.R
\docType{data}
\name{IQncaColors}
\alias{IQncaColors}
\title{Color table for IQnca}
\format{
An object of class \code{character} of length 5000.
}
\usage{
IQncaColors
}
\description{
Color table for IQnca
}
\keyword{datasets}
