% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tools_IQReport.R
\name{IQReport}
\alias{IQReport}
\title{Interface to the IQReport command line interface tool}
\usage{
IQReport(RMDfilePath)
}
\arguments{
\item{RMDfilePath}{Relative path to the RMD file.}
}
\value{
No return value. But the generated DOCX file is available in the
same folder as the RMD file.
}
\description{
Allows to generate Word DOCX files from IQReport specific rmd files.
IQReport is not shipped with IQnca but has to be obtained and installed
separately (http://www.intiquan.com/iqreport/).
}
