% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings_plots.R
\name{IQRggplot}
\alias{IQRggplot}
\title{ggplot functionality implementing IQnca style}
\usage{
IQRggplot(..., fontsize = 12)
}
\arguments{
\item{...}{typical ggplot input arguments}

\item{fontsize}{fontsize used for IQR Tools theme}
}
\value{
ggplot2 object
}
\description{
ggplot functionality implementing IQnca style
}
