% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{IQRinitCompliance}
\alias{IQRinitCompliance}
\title{Initializes the compliance mode}
\usage{
IQRinitCompliance(scriptname)
}
\arguments{
\item{scriptname}{Name of analysis script, generating the outputs
of interest for compliance tracking.}
}
\description{
The compliance mode log file generation requires the global
variable COMPLIANCE_MODE_SCRIPT_NAME to be defined in the global
environment. The user needs to take care that this is the same as
the script file name from his/her analysis script, generating the outputs
of interest for compliance tracking.
}
\seealso{
Other Output & Compliance: 
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputFigure}()},
\code{\link{IQRoutputFile}()},
\code{\link{IQRoutputPDFend}()},
\code{\link{IQRoutputPDFstart}()},
\code{\link{IQRoutputPDF}()},
\code{\link{IQRoutputPNG}()},
\code{\link{IQRoutputRDS}()},
\code{\link{IQRoutputTable}()},
\code{\link{IQRoutputValueTable}()},
\code{\link{IQRoutputXPT}()}
}
\concept{Output & Compliance}
