% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{IQRloadCSVdata}
\alias{IQRloadCSVdata}
\title{Simple auxiliary to load a CSV dataset}
\usage{
IQRloadCSVdata(filename)
}
\arguments{
\item{filename}{Path/filename to the raw or zipped csv dataset. Supports files ending with .csv or .csv.gz.
See \code{\link[data.table:fread]{data.table::fread()}} for details.}
}
\value{
A data.frame with the contents of the CSV file
}
\description{
If a .atr file is present, this will be loaded as well.
A .atr is present if the csv file has been created using the
export function from the IQRdataGENERAL object. Loading the .atr
information will give metadata as attributes about covariates etc.
}
\details{
na.strings=c("."," ","","NA","NaN"),stringsAsFactors=FALSE
}
\seealso{
Other General Data I/O: 
\code{\link{IQRloadSASdata}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputXPT}()},
\code{\link{IQRsaveCSVdata}()}
}
\author{
Henning Schmidt, Daniel Kaschek, Daniel Lill, IntiQuan
}
\concept{General Data I/O}
