% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprogramming_tools.R
\name{IQRloadSASdata}
\alias{IQRloadSASdata}
\title{Loading SAS data}
\usage{
IQRloadSASdata(
  data_file,
  as.data.frame = TRUE,
  noLabels = TRUE,
  replaceComma = TRUE
)
}
\arguments{
\item{data_file}{Path to sas7bdat data file.}

\item{as.data.frame}{Logical. If TRUE a data.frame is returned.}

\item{noLabels}{Logical. If TRUE label attributes are removed}

\item{replaceComma}{If TRUE the "," is exchanged for " " - otherwise not}
}
\value{
data frame
Variable labels are stored in the "label" attribute of each variable. It is
not printed on the console, but the RStudio viewer will show it.
}
\description{
Using the haven package to load SAS data:
Reading supports both sas7bdat and xpt files.
}
\details{
Some arguments from haven::read_sas are disabled on purpose as
there are till bugs with them ... for now only use it for importing SAS files :)

Removing "," in character columns by default (can be switched off)
}
\seealso{
Other General Data I/O: 
\code{\link{IQRloadCSVdata}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputXPT}()},
\code{\link{IQRsaveCSVdata}()}
}
\concept{General Data I/O}
