% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{IQRoutputCSV}
\alias{IQRoutputCSV}
\title{Exporting a data.frame as CSV file}
\usage{
IQRoutputCSV(
  data,
  filename,
  na = ".",
  quote = FALSE,
  row.names = FALSE,
  FLAGattributes = TRUE,
  replaceComma = NULL
)
}
\arguments{
\item{data}{The data.frame or matrix to export}

\item{filename}{Character string with the filename (can include path) for export.
Should include extension to the filename.}

\item{na}{Character to represent NA values}

\item{quote}{TRUE: use "" quotes, FALSE: do not (for NONMEM and MONOLIX)}

\item{row.names}{TRUE: exposrt row names FALSE: do not (typical use in IQnca)}

\item{FLAGattributes}{FALSE: do not export custom attributes as .atr file. TRUE: do (default)}

\item{replaceComma}{NULL: do not replace comma. Character: replace by user-provided character}
}
\description{
The compliance log file will only be generated if .COMPLIANCE_MODE is set to TRUE.
The name of the generated compliance log file is the same name as the text file name
but with appended ".log". The extension for the exported text file will be kept
unchanged in the exported log file name. In order for the log file generation to
work the global variable "COMPLIANCE_MODE_SCRIPT_NAME" needs to be defined by the user
and in each R script set to the name of the script.
}
\seealso{
Other Output & Compliance: 
\code{\link{IQRinitCompliance}()},
\code{\link{IQRoutputFigure}()},
\code{\link{IQRoutputFile}()},
\code{\link{IQRoutputPDFend}()},
\code{\link{IQRoutputPDFstart}()},
\code{\link{IQRoutputPDF}()},
\code{\link{IQRoutputPNG}()},
\code{\link{IQRoutputRDS}()},
\code{\link{IQRoutputTable}()},
\code{\link{IQRoutputValueTable}()},
\code{\link{IQRoutputXPT}()}

Other General Data I/O: 
\code{\link{IQRloadCSVdata}()},
\code{\link{IQRloadSASdata}()},
\code{\link{IQRoutputXPT}()},
\code{\link{IQRsaveCSVdata}()}
}
\author{
Henning Schmidt, Daniel Kaschek, Daniel Lill, IntiQuan
}
\concept{General Data I/O}
\concept{Output & Compliance}
