% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{IQRoutputFile}
\alias{IQRoutputFile}
\title{Exporting a text file with added compliance information log file}
\usage{
IQRoutputFile(text, filename)
}
\arguments{
\item{text}{Character string to export in a file}

\item{filename}{Character string with the filename (can include path) for export.
Should include extension to the filename.}
}
\description{
The compliance log file will only be generated if .COMPLIANCE_MODE is set to TRUE.
The name of the generated compliance log file is the same name as the text file name
but with appended ".log". The extension for the exported text file will be kept
unchanged in the exported log file name. In order for the log file generation to
work the global variable "COMPLIANCE_MODE_SCRIPT_NAME" needs to be defined by the user
and in each R script set to the name of the script.
If compliance model is enabled, the original graphics object will also be stored as
an RDS file, allowing subsequent further processing.
}
\seealso{
Other Output & Compliance: 
\code{\link{IQRinitCompliance}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputFigure}()},
\code{\link{IQRoutputPDFend}()},
\code{\link{IQRoutputPDFstart}()},
\code{\link{IQRoutputPDF}()},
\code{\link{IQRoutputPNG}()},
\code{\link{IQRoutputRDS}()},
\code{\link{IQRoutputTable}()},
\code{\link{IQRoutputValueTable}()},
\code{\link{IQRoutputXPT}()}
}
\concept{Output & Compliance}
