% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{IQRoutputPDFend}
\alias{IQRoutputPDFend}
\title{Ends export to PDF (deprecated)}
\usage{
IQRoutputPDFend(filename)
}
\arguments{
\item{filename}{Name of the PDF file}
}
\description{
PDF export is started with IQRoutputPDFstart. Then all graphical output
goes to this PDF file. PDF export into this file is ended by
IQRoutputPDFend. If the compliance mode is on (.COMPLIANCE_MODE is set to TRUE)
then the exported PDF file is annotated by an additional log file.
The name of the generated compliance log file is the same name as the text file name
but with appended ".log". The extension for the exported text file will be kept
unchanged in the exported log file name. In order for the log file generation to
work the global variable "COMPLIANCE_MODE_SCRIPT_NAME" needs to be defined by the user
and in each R script set to the name of the script.
}
\details{
It is the call to IQRoutputPDFend that actually generates the log file.
}
\seealso{
Other Output & Compliance: 
\code{\link{IQRinitCompliance}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputFigure}()},
\code{\link{IQRoutputFile}()},
\code{\link{IQRoutputPDFstart}()},
\code{\link{IQRoutputPDF}()},
\code{\link{IQRoutputPNG}()},
\code{\link{IQRoutputRDS}()},
\code{\link{IQRoutputTable}()},
\code{\link{IQRoutputValueTable}()},
\code{\link{IQRoutputXPT}()}
}
\concept{Output & Compliance}
