% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{IQRoutputPNG}
\alias{IQRoutputPNG}
\title{Printing a single ggplot object to png file with added compliance information log file}
\usage{
IQRoutputPNG(
  gr,
  filename = NULL,
  res = 300,
  width = 21/2.54,
  height = 21/2.54 * 3/4,
  scale = 1,
  scaleWidth = 1,
  scaleHeight = 1,
  ...
)
}
\arguments{
\item{gr}{ggplot graph object to be printed to file}

\item{filename}{Character string with the filename (can include path) for export.
Should include extension to the filename.}

\item{res}{Resolution in DPI}

\item{width}{width of figure in inch}

\item{height}{height of figure in inch}

\item{scale}{relative factor for scaling of both width and height}

\item{scaleWidth}{relative factor for scaling width}

\item{scaleHeight}{relative factor for scaling height}

\item{...}{whatever Dx put there}
}
\description{
The compliance log file will only be generated if .COMPLIANCE_MODE is set to TRUE.
The name of the generated compliance log file is the same name as the text file name
but with appended ".log". The extension for the exported text file will be kept
unchanged in the exported log file name. In order for the log file generation to
work the global variable "COMPLIANCE_MODE_SCRIPT_NAME" needs to be defined by the user
and in each R script set to the name of the script.
If compliance model is enabled, the original graphics object will also be stored as
an RDS file, allowing subsequent further processing.
}
\seealso{
Other Output & Compliance: 
\code{\link{IQRinitCompliance}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputFigure}()},
\code{\link{IQRoutputFile}()},
\code{\link{IQRoutputPDFend}()},
\code{\link{IQRoutputPDFstart}()},
\code{\link{IQRoutputPDF}()},
\code{\link{IQRoutputRDS}()},
\code{\link{IQRoutputTable}()},
\code{\link{IQRoutputValueTable}()},
\code{\link{IQRoutputXPT}()}
}
\concept{Output & Compliance}
