% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{IQRoutputValueTable}
\alias{IQRoutputValueTable}
\title{Exporting a value with label and potential description for reporting}
\usage{
IQRoutputValueTable(
  value,
  label,
  description = "",
  path = "../Output/00_ValueTable"
)
}
\arguments{
\item{value}{A single value (numeric or string) to report in the table.}

\item{label}{A label allowing to reference the element. Only characters A-Z, a-z, 0-9 are allowed.}

\item{description}{More wordy description of the element. Default: "".}

\item{path}{Path to the folder in which to store the value table elements and where to generate
the value table text file for reporting.}
}
\description{
Single values or elements can be pushed to a table that can be used to
store values of interest during the analysis. Upon call of this
function the path (path argument) an rds file is generated with
the provided label, value, and description content. If compliance
mode is on also the compliance information is stored. In addition
all similar files are read in and a table is constructed and
exported in the text based format used by IQReport.
This table is then useful when reporting results inline in IQReport.
The path to the value table information can be chosen. A default is
provided. The name of the resulting table text file is fixed to
TAB01_ValueTable.txt as there can only be one such file per value
table folder.
}
\seealso{
Other Output & Compliance: 
\code{\link{IQRinitCompliance}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputFigure}()},
\code{\link{IQRoutputFile}()},
\code{\link{IQRoutputPDFend}()},
\code{\link{IQRoutputPDFstart}()},
\code{\link{IQRoutputPDF}()},
\code{\link{IQRoutputPNG}()},
\code{\link{IQRoutputRDS}()},
\code{\link{IQRoutputTable}()},
\code{\link{IQRoutputXPT}()}
}
\concept{Output & Compliance}
