% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input_output.R
\name{IQRsaveCSVdata}
\alias{IQRsaveCSVdata}
\title{Simple auxiliary to save a CSV dataset with attributes}
\usage{
IQRsaveCSVdata(
  data,
  filename,
  na = ".",
  quote = FALSE,
  row.names = FALSE,
  FLAGattributes = TRUE,
  replaceComma = NULL
)
}
\arguments{
\item{data}{A data.frame with the contents of the CSV file}

\item{filename}{Path/filename to the csv dataset.
If file ends with .csv.gz, the csv will be zipped by \code{\link[data.table:fwrite]{data.table::fwrite()}}}

\item{na}{Character to represent NA values}

\item{quote}{TRUE: use "" quotes, FALSE: do not (for NONMEM and MONOLIX)}

\item{row.names}{TRUE: export row names FALSE: do not (typical use in IQnca)}

\item{FLAGattributes}{FALSE: do not export custom attributes as .atr file. TRUE: do (default)}

\item{replaceComma}{NULL: do not replace comma. Character: replace by user-provided character}
}
\description{
If non-standard attributes are present in the data.frame object,
these are stored as a .atr file in addition to the .csv file.
Loading with IQRloadCSVdata will restore the full object with attributes
NA is stored as ".". Compliance mode not supported with this function.
}
\seealso{
Other General Data I/O: 
\code{\link{IQRloadCSVdata}()},
\code{\link{IQRloadSASdata}()},
\code{\link{IQRoutputCSV}()},
\code{\link{IQRoutputXPT}()}
}
\author{
Henning Schmidt, Daniel Kaschek, Daniel Lill, IntiQuan
}
\concept{General Data I/O}
