% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{is_enabled_complianceMode}
\alias{is_enabled_complianceMode}
\title{Checks if the compliance mode is enabled}
\usage{
is_enabled_complianceMode()
}
\value{
Returns TRUE is compliance mode is enabled. FALSE otherwise
}
\description{
Checks if the compliance mode is enabled
}
