% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{is_enabled_rdsMode}
\alias{is_enabled_rdsMode}
\title{Checks if the rds mode is enabled}
\usage{
is_enabled_rdsMode()
}
\value{
Returns TRUE is rds mode is enabled. FALSE otherwise
}
\description{
Checks if the rds mode is enabled
}
