% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{is_IQdataNCA}
\alias{is_IQdataNCA}
\title{Check if object is an IQdataNCA object}
\usage{
is_IQdataNCA(input)
}
\arguments{
\item{input}{object to check}
}
\value{
TRUE if it is an IQdataNCA object. FALSE if it is not
}
\description{
Check if object is an IQdataNCA object
}
