% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{is_IQnca}
\alias{is_IQnca}
\title{Check if object is an IQnca object}
\usage{
is_IQnca(input)
}
\arguments{
\item{input}{object to check}
}
\value{
TRUE if it is an IQnca object. FALSE if it is not
}
\description{
Check if object is an IQnca object
}
