% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{listing_concdetailed_IQdataNCA}
\alias{listing_concdetailed_IQdataNCA}
\title{Generate a detailed listing of individual pharmacokinetic concentrations}
\usage{
listing_concdetailed_IQdataNCA(
  data,
  listingnumber = NULL,
  fontsizetable = 8,
  filename = "listing_concdetailed",
  signifATIME = 5
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{listingnumber}{Character string with listing number information, added to the title for each table, if defined}

\item{fontsizetable}{Fontsize to be used in the table}

\item{filename}{Filename to export the listing IQRrmd object to}

\item{signifATIME}{Significant digits for actual time}
}
\value{
An IQRrmd object exported to a file
}
\description{
The listing will be generated already as an IQRrmd object, allowing easy reporting in Word with IQReport.
The listing is done per USUBJID and provides far more detail than the listing_conc_IQdataNCA function.
Important: Ignored records (INGORER) are not reported in these listings! Ignored subjects (IGNOREI) are reported
but it is indicated that they are ignored (not in the simple version)
}
\seealso{
Other NCA Data Listings: 
\code{\link{listing_conc_IQdataNCA}()},
\code{\link{listing_ignoredSubjectsRecords_IQdataNCA}()},
\code{\link{listing_time_IQdataNCA}()},
\code{\link{listing_timedetailed_IQdataNCA}()}
}
\concept{NCA Data Listings}
