% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{listing_ignoredSubjectsRecords_IQdataNCA}
\alias{listing_ignoredSubjectsRecords_IQdataNCA}
\title{Listing of ignored subjects and records}
\usage{
listing_ignoredSubjectsRecords_IQdataNCA(
  data,
  listingnumber = NULL,
  filename = NULL
)
}
\arguments{
\item{data}{IQdataNCA object for which to do the reporting}

\item{listingnumber}{Character string with listing number information, added to the title for each table, if defined}

\item{filename}{Filename, possibly including path, to store are IQRoutputTable text file for Word reporting}
}
\description{
Reports ignored subjects and records with the reason for ignoring from the analysis.
Displays on screen and can also save the information to a file as an IQRoutputTable object.
}
\seealso{
Other NCA Data Listings: 
\code{\link{listing_conc_IQdataNCA}()},
\code{\link{listing_concdetailed_IQdataNCA}()},
\code{\link{listing_time_IQdataNCA}()},
\code{\link{listing_timedetailed_IQdataNCA}()}
}
\concept{NCA Data Listings}
