% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings_classes_IQdataNCA.R
\name{listing_PKconc_IQdataNCA}
\alias{listing_PKconc_IQdataNCA}
\title{Generate a detailed listing of individual pharmacokinetic concentrations}
\usage{
listing_PKconc_IQdataNCA(
  data,
  table_number = NULL,
  table_split_by = "USUBJID",
  table_head = c("SEX", "AGE", "RACE"),
  table_cols = c("PROFILE", "GROUP", "DOSE"),
  labels = NULL,
  roundfun = signif,
  digits = NULL,
  fontsize = 8,
  page_orientation = c("landscape", "portrait"),
  filename = NULL
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{table_number}{Character string with listing number information, added to the title for each table, if defined}

\item{table_split_by}{Column name to define the split to different tables (pages)}

\item{table_head}{Column names of covariates unique within splits to be displayed above the table}

\item{table_cols}{Column names to be displayed as table columns}

\item{labels}{Named vector with labels that should be used for column names or as name for the common covariates.}

\item{roundfun}{Rounding function or named list of rounding function (expecting the numeric vector and number of digits to be input arguments). Defaults to \link{signif}. Suggested alternative is \link{round}.}

\item{digits}{Named vector with digits to be used for rounding for different columns}

\item{fontsize}{Fontsize to be used in the table}

\item{page_orientation}{Page orientation, 'landscape' or 'portrait'}

\item{filename}{Filename to export the listing IQRrmd object to}
}
\value{
An IQRrmd object exported to a file or IQncaTableList
}
\description{
The listing is split by a user-given column and default to USUBJID.
If a filename is given, an rmd file is generated allowing easy reporting in Word with IQReport. Otherwise,
a IQncaTableList object is returned containing all tables, titles, and footers which can be exported as an rmd file using
\link{write_IQncaTableList}.
Important: Ignored records (INGORER) are not reported in these listings! Ignored subjects (IGNOREI) are reported
but it is indicated that they are ignored (not in the simple version)
}
\seealso{
Other TLF customized: 
\code{\link{figure_indivConc_IQdataNCA}()},
\code{\link{figure_lines_IQdataNCA}()},
\code{\link{figure_summary_IQdataNCA}()},
\code{\link{listing_PKpars_IQnca}()},
\code{\link{table_summary_PKconc_IQdataNCA}()},
\code{\link{table_summary_PKpars_IQnca}()}
}
\concept{TLF customized}
