% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{listing_pkparameter_IQnca}
\alias{listing_pkparameter_IQnca}
\title{Generate a listing of calculated PK parameters}
\usage{
listing_pkparameter_IQnca(
  data,
  listingnumber = NULL,
  fontsizetable = 8,
  filename = "listing_pkparam",
  SIGNIF = 6,
  parameterReport = "standard"
)
}
\arguments{
\item{data}{IQnca object}

\item{listingnumber}{Character string with listing number information, added to the title for each table, if defined}

\item{fontsizetable}{Fontsize to be used in the table}

\item{filename}{Filename to export the listing IQRrmd object to}

\item{SIGNIF}{Number of significant digits}

\item{parameterReport}{Character string defining the PK parameters to report in the listing.
By default these parameters will be selected based on the "Standard" parameters defined in the IQparamNCA_Specification.xlsx
document, subject to the profile and administration type. The user can select "All" or "Standard".
Alternatively, parameterReport can also be a vector or parameter names (based on PKPARAMCD in IQparamNCA_Specification.xlsx).
Instead of the standard or all ones then these will be considered in the listing (also subject to profile and
administration type). Note that interval AUC (if calculated is always added by default!)}
}
\value{
An IQRrmd object exported to a file
}
\description{
The listing will be generated already as an IQRrmd object, allowing easy reporting in Word with IQReport.
The listing is done per USUBJID and PROFILE.
}
\concept{NCA Parameter Listings}
