% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings_classes_IQdataNCA.R
\name{listing_PKpars_IQnca}
\alias{listing_PKpars_IQnca}
\title{Generate a detailed listing of individual pharmacokinetic parameters}
\usage{
listing_PKpars_IQnca(
  data,
  table_number = NULL,
  table_split_by = "USUBJID",
  table_head = c("SEX", "AGE", "RACE"),
  table_cols = c("PROFILE", "GROUP", "DOSE"),
  table_pkpars = "standard",
  pkpars_as_cols = FALSE,
  labels = NULL,
  roundfun = signif,
  digits = 4,
  fontsize = 8,
  page_orientation = c("landscape", "portrait"),
  filename = NULL
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{table_number}{Character string with listing number information, added to the title for each table, if defined}

\item{table_split_by}{Column name to define the split to different tables}

\item{table_head}{Column names of covariates unique within splits to be displayed above the table}

\item{table_cols}{Column names to be displayed as table columns}

\item{table_pkpars}{Character string defining the PK parameters to report in the table
By default these parameters will be selected based on the "Standard" parameters defined in the IQparamNCA_Specification.xlsx
document, subject to the profile and administration type. The user can select "All" or "Standard".
Alternatively, parameterReport can also be a vector or parameter names (based on PKPARAMCD in IQparamNCA_Specification.xlsx).
Instead of the standard or all ones then these will be considered in the listing (also subject to profile and
administration type).}

\item{pkpars_as_cols}{flag whether to display parameters in colums. Defaults to FALSE such that parameters are listed in rows.}

\item{labels}{Named vector with labels that should be used for column names or as name for the common covariates.}

\item{roundfun}{Rounding function (expecting the numeric vector and number of digits to be input arguments). Defaults to \link{signif}. Suggested alternative is \link{round}.}

\item{digits}{Named vector with digits to be used for rounding for different columns}

\item{fontsize}{Fontsize to be used in the table}

\item{page_orientation}{Page orientation, 'landscape' or 'portrait'}

\item{filename}{Filename to export the listing IQRrmd object to}
}
\value{
An IQRrmd object exported to a file or IQncaTableList
}
\description{
The listing is split by a user-given column and defaults to USUBJID.
If a filename is given, an rmd file is generated allowing easy reporting in Word with IQReport. Otherwise,
a IQncaTableList object is returned containing all tables, titles, and footers which can be exported as an rmd file using
\link{write_IQncaTableList}.
}
\details{
As a default, PK parameters are displayed in rows which can be switched to a column-wise table.
}
\examples{
\dontrun{
# Split by USUBJID, one parameter per row
listing_PKpars_IQnca(
  resNCA,
  table_number = "2.1",
  table_split_by = "USUBJID",
  table_cols = c("DOSE", "FORM"),
  table_head = c("USUBJID","SEX", "AGE", "RACE", "SEQUENCE"),
  table_pkpars = c("TMAX", "CMAX", "AUCLST", "AUCTAU", "LAMZHL", "AUCPEP"),
  roundfun = list("TMAX" = round, "CMAX" = round, "AUCLST" = round, "AUCTAU" = round, "LAMZHL" = signif, "AUCPEP" = round),
  digits   = c("TMAX" = Inf, "CMAX" = 2, "AUCLST" = 1, "AUCTAU" = 1, "LAMZHL" = 4, "AUCPEP" = 2),
  filename = file.path(.outputFolder, "01_lising_pars")
)

# Split by PROFILE, parameter as columns
listing_PKpars_IQnca(
  resNCA,
  table_number = "2.1",
  table_split_by = "PROFILE",
  table_cols = "USUBJID",
  table_head = c("FORM","DOSE"),
  table_pkpars = c("TMAX", "CMAX", "AUCLST", "AUCTAU", "LAMZHL", "AUCPEP"),
  pkpars_as_cols = TRUE,
  roundfun = list("TMAX" = round, "CMAX" = round, "AUCLST" = round, "AUCTAU" = round, "LAMZHL" = signif, "AUCPEP" = round),
  digits   = c("TMAX" = Inf, "CMAX" = 2, "AUCLST" = 1, "AUCTAU" = 1, "LAMZHL" = 4, "AUCPEP" = 2),
  filename = file.path(.outputFolder, "02_lising_pars_wide")
)
}
}
\seealso{
Other TLF customized: 
\code{\link{figure_indivConc_IQdataNCA}()},
\code{\link{figure_lines_IQdataNCA}()},
\code{\link{figure_summary_IQdataNCA}()},
\code{\link{listing_PKconc_IQdataNCA}()},
\code{\link{table_summary_PKconc_IQdataNCA}()},
\code{\link{table_summary_PKpars_IQnca}()}
}
\concept{TLF customized}
