% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{listing_time_IQdataNCA}
\alias{listing_time_IQdataNCA}
\title{Generate a listing of individual pharmacokinetic concentration sampling times}
\usage{
listing_time_IQdataNCA(
  data,
  listingnumber = NULL,
  fontsizetable = 8,
  filename = "listing_time",
  signifATIME = 5,
  maxCol = 10
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{listingnumber}{Character string with listing number information, added to the title for each table, if defined}

\item{fontsizetable}{Fontsize to be used in the table}

\item{filename}{Filename to export the listing IQRrmd object to}

\item{signifATIME}{Significant digits for actual time}

\item{maxCol}{Number of nominal times to be put into the same table in case of simple=TRUE. Tables might become to wide even for landscape
so a plit up of columns in several tables can be enforced in this manner.}
}
\value{
An IQRrmd object exported to a file
}
\description{
The listing will be generated already as an IQRrmd object, allowing easy reporting in Word with IQReport.
It is a simple listing - just showing ATIME and NTIME by PROFILE and GROUP.
Important: Ignored records (INGORER) are not reported in these listings! Ignored subjects (IGNOREI) are reported
but it is indicated that they are ignored (not in the simple version).
}
\seealso{
Other NCA Data Listings: 
\code{\link{listing_conc_IQdataNCA}()},
\code{\link{listing_concdetailed_IQdataNCA}()},
\code{\link{listing_ignoredSubjectsRecords_IQdataNCA}()},
\code{\link{listing_timedetailed_IQdataNCA}()}
}
\concept{NCA Data Listings}
