% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{listing_timedetailed_IQdataNCA}
\alias{listing_timedetailed_IQdataNCA}
\title{Generate a listing of individual pharmacokinetic concentration sampling times}
\usage{
listing_timedetailed_IQdataNCA(
  data,
  listingnumber = NULL,
  fontsizetable = 8,
  filename = "listing_timedetailed",
  signifATIME = 5
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{listingnumber}{Character string with listing number information, added to the title for each table, if defined}

\item{fontsizetable}{Fontsize to be used in the table}

\item{filename}{Filename to export the listing IQRrmd object to}

\item{signifATIME}{Significant digits for actual time}
}
\value{
An IQRrmd object exported to a file
}
\description{
The listing will be generated already as an IQRrmd object, allowing easy reporting in Word with IQReport.
The listing is done per USUBJID and contains more information than the listing_time_IQdataNCA result.
Important: Ignored records (INGORER) are not reported in these listings! Ignored subjects (IGNOREI) are reported
but it is indicated that they are ignored (not in the simple version).
}
\seealso{
Other NCA Data Listings: 
\code{\link{listing_conc_IQdataNCA}()},
\code{\link{listing_concdetailed_IQdataNCA}()},
\code{\link{listing_ignoredSubjectsRecords_IQdataNCA}()},
\code{\link{listing_time_IQdataNCA}()}
}
\concept{NCA Data Listings}
