% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{loadSetupOptions_IQnca}
\alias{loadSetupOptions_IQnca}
\title{Loading IQnca setup options}
\usage{
loadSetupOptions_IQnca()
}
\description{
These options are stored in the installed R package as the file
"setup_options_IQnca.R". Typically the sysadmin or the user itself can
edit this file. In addition, the user can have an own user setup_options_IQnca.R
file located in the users home folder. If present then the contents of this
file are loaded after the contents of the general file, allowing the user
to override some of the general settings.
}
\details{
Note that while the contents of the general file might change from installed
version to version, the contents of the file in the users home folder are
not changed. Updates to this file need to be done by the user.
}
