% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA_DK.R
\name{mutate_IQnca}
\alias{mutate_IQnca}
\title{Modify or add columns (PK parameters) to the NCA result}
\usage{
mutate_IQnca(
  data,
  name,
  based_on = NULL,
  formula,
  name_short = NULL,
  name_long = NULL,
  description = NULL,
  unit = NULL,
  sd = NULL,
  fd = NULL,
  ss = NULL,
  bolus = NULL,
  infusion = NULL,
  extravascular = NULL,
  reporting = NULL,
  reporting_across_dose = NULL
)
}
\arguments{
\item{data}{IQnca object, i.e., the output of \link{nca_IQdataNCA}.}

\item{name}{Character of length 1. The column name to be generated (corresponds to PKPARAMCD).}

\item{based_on}{Character of length 1 or NULL (default). An existing column from which the new
parameter inherits all the attributes which are not explicitly provided by other argument of
this function.}

\item{formula}{One-sided formula of the form "~expression". Here expression is an algebraic
expression by which the new parameter is computed based on existing columns in \code{data}. See
details for more information.}

\item{name_short}{Character of length 1 or NULL (default). A short, humanly readable name of
the added parameter. If \code{NULL}, the value proposed by \code{based_on} will be used. If also \code{based_on}
is \code{NULL}, \code{name} will be used.}

\item{name_long}{Character of length 1 or NULL (default). A slightly longer, humanly readable
name of the added parameter. \strong{This name will appear in tables and listings}. If \code{NULL}, the
value proposed by \code{based_on} will be used. If also \code{based_on} is \code{NULL}, \code{name} will be used.}

\item{description}{Character of length 1 or NULL (default). A longer description of
the added parameter. If \code{NULL}, the value proposed by \code{based_on} will be used. If also \code{based_on}
is \code{NULL}, \code{name} will be used.}

\item{unit}{Character of length 1 or NULL (default). The unit can be provided explicitly,
e.g., mg, mL, kg, etc., or using keywords recognized by IQnca, see details. If \code{NULL}, the
value proposed by \code{based_on} will be used. If also \code{based_on} is \code{NULL}, "-" will be used.}

\item{sd}{Logical or NULL (default). Indicates whether the param should be reported for
single dose profiles. If \code{NULL}, the value proposed by \code{based_on} will be used. If also
\code{based_on} is \code{NULL}, \code{TRUE} will be assumed.}

\item{fd}{Logical or NULL (default). Indicates whether the param should be reported for
first dose profiles. If \code{NULL}, the value proposed by \code{based_on} will be used. If also
\code{based_on} is \code{NULL}, \code{TRUE} will be assumed.}

\item{ss}{Logical or NULL (default). Indicates whether the param should be reported for
steady-state profiles. If \code{NULL}, the value proposed by \code{based_on} will be used. If also
\code{based_on} is \code{NULL}, \code{TRUE} will be assumed.}

\item{bolus}{Logical or NULL (default). Indicates whether the param should be reported for
administration type "BOLUS". If \code{NULL}, the value proposed by \code{based_on} will be used. If also
\code{based_on} is \code{NULL}, \code{TRUE} will be assumed.}

\item{infusion}{Logical or NULL (default). Indicates whether the param should be reported for
administration type "INFUSION". If \code{NULL}, the value proposed by \code{based_on} will be used. If also
\code{based_on} is \code{NULL}, \code{TRUE} will be assumed.}

\item{extravascular}{Logical or NULL (default). Indicates whether the param should be reported for
administration type "EXTRAVASCULAR". If \code{NULL}, the value proposed by \code{based_on} will be used. If also
\code{based_on} is \code{NULL}, \code{TRUE} will be assumed.}

\item{reporting}{"standard", "slope", "other", or \code{NULL} (default). See details. If \code{NULL}, the value
proposed by \code{based_on} will be used. If also \code{based_on} is \code{NULL}, "standard" will be assumed.}

\item{reporting_across_dose}{Logical or NULL (default). Indicates whether the parameter should
appear in summary tables for which PK parameters were summarized across doses. See details.
If \code{NULL}, the value proposed by \code{based_on} will be used. If also \code{based_on} is \code{NULL},
\code{FALSE} will be assumed.}
}
\value{
IQnca object including the new or modified column, an updated units attribute and
an updated parameter specification table (paramspec) attribute.
}
\description{
The function allows to overwrite or derive new PK parameters based on the existing parameters,
e.g., transforming the result to a different unit, or deriving per body weight parameters.
Alternatively, if the computation of additional PK parameters requires more than
just an algebraic relationship, the function can be used to "register" manually
added columns of custom PK parameters.
}
\details{
For the interpretation of \strong{\code{formula}}, the right-hand side of the formula is parsed
to an R \code{expression} which is then evaluated in a local environment generated by \code{data} using
the \code{with} function.

\strong{Units} can be stated explicitly or using key words recognized by IQnca. Valid keywords
are \code{CONCUNIT}, \code{TIMEUNIT}, \code{DOSEUNIT}, \code{CLROUTINE}, and \code{VOLROUTINE}. Here, \code{CLROUTINE}
and \code{VOLROUTINE} refer to clearance and volume routines. These units are computed by
internal routines.

\strong{Reporting} of values is categorized by "standard", "slope", and "other". Here, "standard"
indicates parameters which are reported by default, "slope" indicates parameters related
to slope determination, and "other" are the remaining PK parameters. Which parameters
are acutally reported in tables and listings can be fine-tuned by the table and listings
functions if IQnca.

\strong{Reporting across dose} means that PK parameters summarized across different dose groups
are reported. Summarizing across doses can be valid for parameters which are expected
to be independent of dose, e.g., TMAX, clearance, terminal half-life. Summarizing across
doses is usually not valid for parameters such as Cmax and AUC.
}
\seealso{
Other NCA Parameter Tables: 
\code{\link{reliabilityTable_IQnca}()},
\code{\link{table_summary_acrossGROUP_pkparameters_IQnca}()},
\code{\link{table_summary_byGROUP_pkparameters_IQnca}()}

Other NCA Parameter Listings: 
\code{\link{listing_pkparameter_IQnca}()}

Other NCA Parameter Export: 
\code{\link{exportCDISC_IQnca}()},
\code{\link{export_IQnca}()}
}
\author{
Daniel Kaschek, IntiQuan
}
\concept{NCA Parameter Export}
\concept{NCA Parameter Listings}
\concept{NCA Parameter Tables}
