% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{nca_IQdataNCA}
\alias{nca_IQdataNCA}
\title{Calculation of NCA PK parameters}
\usage{
nca_IQdataNCA(data, milliLiter = FALSE, CTLASTwinnonlinbehavior = TRUE)
}
\arguments{
\item{data}{IQdataNCA object}

\item{milliLiter}{By default L (Liter) is used as a standard unit for volume. If desired, by setting this
flag to TRUE also mL (milliliter) can be used. This is defined here during the calculation of the
PK parameters and not in the dataset as numerically it will give the same results (with some commata shifts).}

\item{CTLASTwinnonlinbehavior}{If BLLOQ handling rule is set to use LLOQ/2 or LLOQ
for the first BLLOQ value post last observable, then Clast and Tlast will be set based
on this first BLLOQ imputation. This is counterintuitive - but in order to be aligned
with Winnonlin we had to add this exception. We do so here as an option ...
Use TRUE by default.}
}
\value{
IQnca object with calculated parameters. IQdataNCA object is added as attribute "dataNCA".
Information about column names and labels for interval AUC is also stored as attributes.
}
\description{
Ignored: ignored subjects (IGNOREI), ignored records (IGNORER and IGNORENCA)
Computed parameters and algorithm settings fully controlled by settings in the
IQdataNCA object.
}
