% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{opt.layout}
\alias{opt.layout}
\title{Convenience function for setting additional
page layout options (opt.layout) for IQRoutputFigure}
\usage{
opt.layout(
  nrow = NULL,
  ncol = NULL,
  npage = NULL,
  legend.option = NULL,
  legend.object = NULL,
  legend.position = NULL,
  legend.relsize = NULL,
  title.relheight = NULL,
  subtitle.relheight = NULL,
  footer.relheight = NULL
)
}
\arguments{
\item{nrow}{Number of rows per page}

\item{ncol}{Number of columns per page}

\item{npage}{Number of pages (ignored if both nrow and ncol are given)}

\item{legend.option}{Character or numeric whether to leave legend in plots as is ('as.is', 1),
remove all legends from the plots ('remove', 2), or plot common legend ('common', 3).
As common legend, the legend of the first plot per page is used if not user provided
by 'legend.object'}

\item{legend.object}{User-provided legend}

\item{legend.position}{Position of legend relative to main plots.
Can be 'right' (default), 'left', 'top', or 'bottom'.}

\item{legend.relsize}{Fraction of plot region (width for left or right position,
height for top or bottom position) allocated to legend.}

\item{title.relheight}{Fraction of page height allocated to plot title (if exists).}

\item{subtitle.relheight}{Fraction of page height allocated to plot subtitle (if exists).}

\item{footer.relheight}{Fraction of page height allocated to plot footer (if exists).}
}
\value{
Returns a opt.layout list for IQRoutputFigure
}
\description{
Convenience function for setting additional
page layout options (opt.layout) for IQRoutputFigure
}
\author{
Anne Kuemmel, IntiQuan
}
