% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{opt.pagesize}
\alias{opt.pagesize}
\title{Convenience function for setting additional
page printing options (opt.pagesize) for IQRoutputFigure}
\usage{
opt.pagesize(
  width = NULL,
  height = NULL,
  scale = NULL,
  scaleWidth = NULL,
  scaleHeight = NULL,
  res = NULL
)
}
\arguments{
\item{width}{Page with in inches}

\item{height}{Page height in inches}

\item{scale}{Relative factor for scaling of both width and height}

\item{scaleWidth}{Relative factor for scaling width}

\item{scaleHeight}{Relative factor for scaling height}

\item{res}{Resolution in DPI (applies for PNG output)}
}
\value{
Returns a opt.pagesize list for IQRoutputFigure
}
\description{
Convenience function for setting additional
page printing options (opt.pagesize) for IQRoutputFigure
}
\author{
Anne Kuemmel, IntiQuan
}
