% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{ordercol_IQdataNCA}
\alias{ordercol_IQdataNCA}
\title{Order an IQdataNCA objects columns in standard way}
\usage{
ordercol_IQdataNCA(data)
}
\arguments{
\item{data}{data.frame for which the columns should be ordered}
}
\value{
data.frame with order of columns adjusted
}
\description{
Works also when not all columns are present yet. Same order used as defined based on the list and entries
returned by function getColumnNames_IQdataNCA(). Columns that are present in the data.frame but not
named in getColumnNames_IQdataNCA() will be added at the end in any order
}
