% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{pcorres2aconc_IQdataNCA}
\alias{pcorres2aconc_IQdataNCA}
\title{Convert typical PCORRES column in PC domain to AVALUE col in IQdataNCA}
\usage{
pcorres2aconc_IQdataNCA(PCORRES)
}
\arguments{
\item{PCORRES}{Character vector with "BLQ" for BLLOQ data, other alphanumerical entries point to unavailable measurements, and
values for actual measurements}
}
\value{
Numeric vector
}
\description{
Assumption: PCORRES uses "BLQ" for BLLOQ values and all other alphanumerical entries point to unavailable measurements
}
