% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{plot.IQdataNCA}
\alias{plot.IQdataNCA}
\title{Generic plot function for IQdataNCA objects}
\usage{
\method{plot}{IQdataNCA}(x, ..., logY = TRUE, filename = NULL, nindiv = 4, R2ADJTHRESHOL = 0.85)
}
\arguments{
\item{x}{IQdataNCA object}

\item{...}{Whatever R wants here. These dots are useless and annoying}

\item{logY}{Log y axis is TRUE, otherwise linear Y axis}

\item{filename}{Name of PDF file to generate with plots}

\item{nindiv}{Number of plots per page if exported to PDF}

\item{R2ADJTHRESHOL}{Threshold for R2ADJ to switch color and annotation
to warn when R2ADJ<R2ADJTHRESHOL}
}
\value{
A list with individual plots
}
\description{
The purpose is to allow exploratory plots of the data. The goal is not
to produce submission ready figures.
}
\details{
Plotting single individual (USUBJID)
TIME / CONCPLIN or CONCPLOT
Adding additional meta information, including slope calculation, etc.
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()}
}
\concept{NCA Data}
