% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{plot.IQRoutputFigure}
\alias{plot.IQRoutputFigure}
\title{Plot function for IQRoutputFigure (calls print)}
\usage{
\method{plot}{IQRoutputFigure}(x, ...)
}
\arguments{
\item{x}{IQRoutputFigure object}

\item{...}{Arguments passed to print function}
}
\description{
Plot function for IQRoutputFigure (calls print)
}
\author{
Anne Kümmel, IntiQuan
}
