% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_classes_IQdataNCA.R
\name{plot_lines_IQdataNCA}
\alias{plot_lines_IQdataNCA}
\title{Plot individual concentration data of IQnca data object}
\usage{
plot_lines_IQdataNCA(
  d,
  figure_head,
  lines_by,
  panel_by,
  labels,
  yscale = "linlog",
  dosenorm = FALSE,
  legend_pos,
  xlabtext,
  xlim,
  ylabtext,
  ylimlin,
  ylimlog
)
}
\arguments{
\item{d}{IQncaData object with the subset of data to be plotted here}

\item{figure_head}{Column names of covariates unique within pages to be displayed above the table}

\item{lines_by}{grouping used for linetypes}

\item{panel_by}{variable to create panels by}

\item{labels}{names character vector with labels for columns}

\item{yscale}{character identifying whether to plot linear ("lin")
or semilogarithmic ("log") or both views ("linlog")}

\item{dosenorm}{logical whether to plot dose normalized concentrations}

\item{legend_pos}{specifying legend position}

\item{xlabtext}{label of x axis (time)}

\item{xlim}{limits of x axis (time) to be applied to all panels/pages}

\item{ylabtext}{Label for y axis}

\item{ylimlin}{Limits to be applied to linear y axis. If NULL, no scale enforced.}

\item{ylimlog}{Limits to be applied to logarithmic y axis. If NULL, no scale enforced.}
}
\value{
individual line plots of concentrations
}
\description{
Plot individual concentration data of IQnca data object
}
