% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_classes_IQdataNCA.R
\name{plot_summary_IQdataNCAsum}
\alias{plot_summary_IQdataNCAsum}
\title{Generate one page with summary figure}
\usage{
plot_summary_IQdataNCAsum(
  d,
  lines_by,
  panel_by,
  figure_head,
  labels,
  yscale,
  dosenorm,
  legend_pos,
  xlabtext,
  xlim,
  ylabtext,
  ylimlin,
  ylimlog
)
}
\arguments{
\item{d}{Data frame with summary of PK concentration to be plot on one page}

\item{lines_by}{grouping used for linetypes}

\item{panel_by}{variable to create panels by}

\item{figure_head}{Column names of covariates unique within pages to be displayed above the table}

\item{labels}{names character vector with labels for columns}

\item{yscale}{character identifying whether to plot linear ("lin")
or semilogarithmic ("log") or both views ("linlog")}

\item{dosenorm}{logical whether to plot dose normalized concentrations}

\item{legend_pos}{specifying legend position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{xlabtext}{label of x axis (time)}

\item{xlim}{limits of x axis (time) to be applied to all panels/pages}

\item{ylabtext}{Label for y axis}

\item{ylimlin}{Limits to be applied to linear y axis. If NULL, no scale enforced.}

\item{ylimlog}{Limits to be applied to logarithmic y axis. If NULL, no scale enforced.}
}
\value{
Returns a figure (ggplot or arranged ggplot object)
}
\description{
Generate one page with summary figure
}
\seealso{
Other NCA Data Figures: 
\code{\link{figure_spaghetti_IQdataNCA}()},
\code{\link{figure_spaghetti_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_geomean_IQdataNCA}()},
\code{\link{figure_summary_geommean_dosenorm_IQdataNCA}()},
\code{\link{figure_summary_mean_IQdataNCA}()},
\code{\link{figure_summary_mean_dosenorm_IQdataNCA}()},
\code{\link{figures_indiv_IQdataNCA}()}
}
\concept{NCA Data Figures}
