% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{+.IQRrmd}
\alias{+.IQRrmd}
\title{Concatenate IQRrmd objects}
\usage{
\method{+}{IQRrmd}(a, b)
}
\arguments{
\item{a}{Character string or number or IQRrmd object}

\item{b}{Character string or number or IQRrmd object}
}
\value{
Concatenated IQRrmd object "ab"
}
\description{
Concatenate IQRrmd objects
}
