% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{print.IQdataNCA}
\alias{print.IQdataNCA}
\title{Generic print function for IQdataNCA objects}
\usage{
\method{print}{IQdataNCA}(x, ...)
}
\arguments{
\item{x}{IQdataNCA object}

\item{...}{Additional arguments - whatever R hides in these ... bad style!}
}
\description{
Generic print function for IQdataNCA objects
}
