% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{print.IQnca}
\alias{print.IQnca}
\title{Generic print function for IQnca objects}
\usage{
\method{print}{IQnca}(x, ...)
}
\arguments{
\item{x}{IQnca object}

\item{...}{Additional arguments - whatever R hides in these ... bad style!}
}
\description{
Generic print function for IQnca objects
}
