% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{print.IQRoutputFigure}
\alias{print.IQRoutputFigure}
\title{Print IQRoutputFigure to pages (ggplot objects)}
\usage{
\method{print}{IQRoutputFigure}(
  x,
  nrow = NULL,
  ncol = NULL,
  npage = NULL,
  legend.option = "as.is",
  legend.object = NULL,
  legend.position = "right",
  legend.relsize = 0.2,
  title.relheight = 0.05,
  subtitle.relheight = 0.05,
  footer.relheight = 0.05,
  ...
)
}
\arguments{
\item{x}{IQRoutputFigure object}

\item{nrow}{number of rows per page}

\item{ncol}{number of columns per page}

\item{npage}{number of pages. If nrow and ncol is given, this input is ignored.}

\item{legend.option}{Character or numeric whether to leave legend in plots as is ('as.is', 1),
remove all legends from the plots ('remove', 2), or plot common legend ('common', 3).
As common legend, the legend of the first plot per page is used if not user provided
by 'legend.object'}

\item{legend.object}{User-provided legend}

\item{legend.position}{Position of legend relative to main plots.
Can be 'right' (default), 'left', 'top', or 'bottom'.}

\item{legend.relsize}{Fraction of plot region (width for left or right position,
height for top or bottom position) allocated to legend.}

\item{title.relheight}{Fraction of page height allocated to plot title (if exists).}

\item{subtitle.relheight}{Fraction of page height allocated to plot subtitle (if exists).}

\item{footer.relheight}{Fraction of page height allocated to plot footer (if exists).}

\item{...}{Additional arguments (unused)}
}
\description{
Print IQRoutputFigure to pages (ggplot objects)
}
\author{
Anne Kümmel, IntiQuan
}
