% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{print.IQRoutputFigureList}
\alias{print.IQRoutputFigureList}
\title{Prints summary for the IQRoutputFigures contained in the list}
\usage{
\method{print}{IQRoutputFigureList}(x, ...)
}
\arguments{
\item{x}{IQRplotMulti object}

\item{...}{Additional arguments (unused)}
}
\value{
list of list of ggplot objects, one per page
}
\description{
Prints summary for the IQRoutputFigures contained in the list
}
\author{
Anne Kümmel, IntiQuan
}
