% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{print.IQRoutputTable}
\alias{print.IQRoutputTable}
\title{Overloading "print" for IQRoutputTable objects}
\usage{
\method{print}{IQRoutputTable}(x, ...)
}
\arguments{
\item{x}{IQRoutputTable object}

\item{...}{Additional parameters}
}
\description{
Overloading "print" for IQRoutputTable objects
}
