% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{print.IQRrmd}
\alias{print.IQRrmd}
\title{Overloading "print" for IQRrmd objects}
\usage{
\method{print}{IQRrmd}(x, ...)
}
\arguments{
\item{x}{IQRrmd object}

\item{...}{Additional parameters}
}
\description{
Overloading "print" for IQRrmd objects
}
