% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{printGrid}
\alias{printGrid}
\title{Plotting a list of ggplot objects in a grid defined by nrow and ncol}
\usage{
printGrid(plotList, nrow = 4, ncol = 4)
}
\arguments{
\item{plotList}{List of ggplot objectsx}

\item{nrow}{Number of rows per page}

\item{ncol}{Number of cols per page}
}
\description{
Useful when having a lot of plots that should be presented in a page saving manner.
}
