% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_classes_IQdataNCA.R
\name{rawtable_summary_PKconc_IQdataNCA}
\alias{rawtable_summary_PKconc_IQdataNCA}
\title{Generate a single split of statistical summaries of pharmacokinetic concentrations}
\usage{
rawtable_summary_PKconc_IQdataNCA(
  d,
  stratify_by,
  statistics,
  table_number = NULL,
  split_index,
  table_split_by,
  table_head,
  table_cols,
  table_compare,
  labels,
  NTall,
  roundfun = signif,
  digits,
  filename
)
}
\arguments{
\item{d}{Split of IQdataNCA object}

\item{stratify_by}{Columns that define the groups that are summarized}

\item{statistics}{Statistics to be calculated}

\item{table_number}{Character string with table number information, added to the title for each table, if defined}

\item{split_index}{Indec of the current split}

\item{table_split_by}{Column that the split was done by (for labelling)}

\item{table_head}{Common information for split to be displayed above table}

\item{table_cols}{Columns to be displayed before the summary statistics}

\item{table_compare}{Single column defining groups for which the statistics are displayed side-by-side}

\item{labels}{Customized labels for columns that substitute the column name in the IQdataNCA object. Applies to 'table_head', 'table_cols', and the entries of 'table_compare'}

\item{NTall}{Number of nominal time points}

\item{roundfun}{Rounding function to be used (defaults to \link{signif})}

\item{digits}{Number of digits to be used for rounding}

\item{filename}{Filename to export the table IQRrmd object to}
}
\value{
IQncaTable object
}
\description{
Generate a single split of statistical summaries of pharmacokinetic concentrations
}
