% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA_DK.R
\name{reliabilityTable_IQnca}
\alias{reliabilityTable_IQnca}
\title{Determine reliability of the NCA results from IQnca object}
\usage{
reliabilityTable_IQnca(
  data,
  crit_AUCPEO = 20,
  crit_R2ADJ = 0.85,
  crit_tinterval_thalf = 1.5
)
}
\arguments{
\item{data}{IQnca object.}

\item{crit_AUCPEO}{max. percent AUC extrapolation acceptable, 20\% by default.}

\item{crit_R2ADJ}{min. adjusted Rsquared value required, 0.85 by default.}

\item{crit_tinterval_thalf}{min. length of the time interval from which the terminal
slope was determined, 1.5*thalf by default.}
}
\value{
data.frame with USUBJID, GROUP, PROFILE, the success of each criterion,
and TOTAL = success of all criteria
}
\description{
Reliability is determined based on 3 criteria applying to \%AUC extrapolated,
adjusted Rsquared threshold, and length of the time interval from which the half life
was determined.
}
\seealso{
Other NCA Parameter Tables: 
\code{\link{mutate_IQnca}()},
\code{\link{table_summary_acrossGROUP_pkparameters_IQnca}()},
\code{\link{table_summary_byGROUP_pkparameters_IQnca}()}
}
\concept{NCA Parameter Tables}
