% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Reduce character to the non-repeating entries}
\usage{
remove_duplicates(x, group = NULL)
}
\arguments{
\item{x}{character vector.}

\item{group}{optional vector of the same lenght as x that can be coerced to a factor.
If \code{group} is provided, duplicates are only removed within the same group.}
}
\value{
character of the same length as x.
}
\description{
Entries which are the same as the preceding one are are replaced by empty character.
The inverse of insert_duplicates().
}
