% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprogramming_tools.R
\name{removeCommata_dataframe}
\alias{removeCommata_dataframe}
\title{Remove commata in elements in data.frame}
\usage{
removeCommata_dataframe(data, replaceComma = " ")
}
\arguments{
\item{data}{data.frame with potentially commata in elements}

\item{replaceComma}{Character to replace commata with}
}
\value{
data.frame without commata in elements
}
\description{
Remove commata in elements in data.frame
}
