% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_auxilliary.R
\name{replace_colname_by_label}
\alias{replace_colname_by_label}
\title{Replace the column name by its label if defined.}
\usage{
replace_colname_by_label(df)
}
\arguments{
\item{df}{data frame with potentially labeled columns}
}
\value{
data frame with substituted column names
}
\description{
Applies to the labeled columns only. Others keep there name.
}
\examples{
\dontrun{
df <- replace_colname_by_label(df)
}
}
