% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{report_IQnca}
\alias{report_IQnca}
\title{Generate default NCA report}
\usage{
report_IQnca(
  reportfile,
  resultpaths,
  templatestyle = "DefaultStyle.rmdt",
  title = "Some NCA",
  docx = TRUE
)
}
\arguments{
\item{reportfile}{Report file name. Note that a certain structure is assumed in the generation of the
RMD files containing figures. The location of the report file needs to match this in terms of the
relative paths used in the figure RMD files.}

\item{resultpaths}{Paths to folders with generated RMD results (listings, tables, and figures). All
present RMD files will be included.}

\item{templatestyle}{IQReport template file.}

\item{title}{Title of the report}

\item{docx}{logical. If TRUE and IQReport is present a Word DOCX file will be generated. If FALSE then only
a combined RMD file will be generated. The latter is useful if one wants to continue editing the file as a
stand alone NCA report.}
}
\description{
Collects all generated results (RMD) files and creates a combined RMD file. Optionally,
if IQReport is available it also generates a Word file.
}
\concept{NCA Report}
