% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdAPPROVALS}
\alias{rmdAPPROVALS}
\title{Add an approvals page to the report (after title)}
\usage{
rmdAPPROVALS()
}
\value{
APPROVALS command as IQRrmd object
}
\description{
This command is only working as expected if the used template actually has an
approvals page included. Otherwise this command will be ignored.
}
