% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdCOMMENT}
\alias{rmdCOMMENT}
\title{Generate RMD command to include a Word comment}
\usage{
rmdCOMMENT(comment = "some comment", author = NULL)
}
\arguments{
\item{comment}{Character string, specifying the comment text. Line breaks are
allowed.}

\item{author}{Character string, specifying the author}
}
\value{
COM command as IQRrmd object
}
\description{
Generate RMD command to include a Word comment
}
