% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdEMPTY}
\alias{rmdEMPTY}
\title{Create empty IQRrmd object}
\usage{
rmdEMPTY()

rmdEMPTY()
}
\value{
Empty IQRrmd object

Empty IQRrmd object
}
\description{
Create empty IQRrmd object

Create empty IQRrmd object
}
