% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdFIGURE}
\alias{rmdFIGURE}
\title{Generate RMD command to include a figure from a file}
\usage{
rmdFIGURE(
  file,
  caption = NULL,
  label = NULL,
  legend = NULL,
  pages = NULL,
  scale = 100,
  crop = TRUE,
  repeatLegend = TRUE,
  ignoreCaption = FALSE
)
}
\arguments{
\item{file}{Character string, specifying the absolute or relative path to the
figure file to include (pdf, png, jpg, etc.).}

\item{caption}{Caption for the table. If NULL then the title information in the
table text file is used as caption.}

\item{label}{Character string to be used for cross-referencing.}

\item{legend}{Character string with text that is written as "legend" under the figure.}

\item{pages}{A character string with information about which pages of a multi-page PDF
should be included into the report. If undefined all pages will be exported with the same
caption and figure number. The legend will be repeated after each figure if argument
\code{repeatLegend} is set to TRUE.
Example: pages="3 5-8" exports page 3,5,6,7,and 8.}

\item{scale}{Number value (percent) for scaling the figure.}

\item{crop}{Boolean. If set to TRUE then figure will be cropped to remove surrounding white space.}

\item{repeatLegend}{See \code{legend}.}

\item{ignoreCaption}{Caption can be ignored if desired (no caption will be shown for figure)}
}
\value{
FIG command as IQRrmd object
}
\description{
Allows to generate RMD code to include a figure defined in a pdf, png, jpg file.
From multi page PDFs selected pages or page ranges can be included.
}
