% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdLANDSCAPE}
\alias{rmdLANDSCAPE}
\title{Switch to Landscape mode}
\usage{
rmdLANDSCAPE()
}
\value{
Landscape command as IQRrmd object
}
\description{
This will only work if the section where the landcape command is called contains
some text ... tables or figures do not count ...
}
