% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdNOINTRO}
\alias{rmdNOINTRO}
\title{Do not include TOC/TOF/TOT in DOCX  report}
\usage{
rmdNOINTRO()
}
\value{
NOINTRO command as IQRrmd object
}
\description{
The !NOINTRO command will remove the table of contents, figures, and tables
from the Word document.
}
