% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdPDF}
\alias{rmdPDF}
\title{Generate RMD command to include a PDF from a file}
\usage{
rmdPDF(file, pages = NULL, scale = 100, crop = TRUE)
}
\arguments{
\item{file}{Character string, specifying the absolute or relative path to the
PDF file to include.}

\item{pages}{A character string with information about which pages of a multi-page PDF
should be included into the report.
Example: pages="3 5-8" exports page 3,5,6,7,and 8.}

\item{scale}{Number value (percent) for scaling the figure.}

\item{crop}{Boolean. If set to TRUE then figure will be cropped to remove surrounding white space.}
}
\value{
PDF command as IQRrmd object
}
\description{
The main difference to rmdFIGURE is that rmdPDF does not create a caption and
not an entry into the List of Figures. This allows to include PDF documents as
sections ... if desired. Compliance text is not shown.
}
