% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdSECTION}
\alias{rmdSECTION}
\title{Generate Section RMD command}
\usage{
rmdSECTION(title = "Section Title", label = NULL, level = 1, numbered = TRUE)
}
\arguments{
\item{title}{Character string with title of section}

\item{label}{Character string (no spaces) with label for cross-reference use.
Limited to 40 characters.}

\item{level}{Integer value (1,2,3,etc.) to indicate the section level}

\item{numbered}{Boolean to define if section shoud be numbered or not (experimental)}
}
\value{
Section RMD command as IQRrmd object
}
\description{
Allows generation of RMD strings to implement sections (level=1), subsections
(level=2), subsubsections (level=3), etc. Sections can be numbered (numbered=TRUE)
or unnumbered (numbered=FALSE). A label can be given to be able to crossreference
this section.
}
\details{
Unnumbered sections have the limitation to level=1 and only a single unnumbered section
can be present in a document - the first one.
}
