% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTABLE}
\alias{rmdTABLE}
\title{Generate RMD command to include a table from a file}
\usage{
rmdTABLE(
  file,
  caption = NULL,
  fontsize = 10,
  label = NULL,
  ignoreCaption = FALSE,
  valueTable = FALSE
)
}
\arguments{
\item{file}{Character string, specifying the absolute or relative path to the
table text file to include.}

\item{caption}{Caption for the table. If NULL then the title information in the
table text file is used as caption.}

\item{fontsize}{Integer value for the fontsize.}

\item{label}{Character string to be used for cross-referencing.}

\item{ignoreCaption}{Caption can be ignored if desired (no caption will be shown for table)}

\item{valueTable}{Defines is table can be used as a "Value Table", allowing to reference
values in the table. A value table requires the presence of a "value" and a "label" column.
If these are not present, an error will occur.}
}
\description{
Allows to generate RMD code to include a table defined in a text file.
To include a table from an R dataframe, please use the function
rmdTABLEDF.
}
