% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTABLEDF}
\alias{rmdTABLEDF}
\title{Generate RMD command to include a table from a dataframe}
\usage{
rmdTABLEDF(
  df,
  caption = NULL,
  fontsize = 10,
  label = NULL,
  ignoreCaption = FALSE,
  valueTable = FALSE,
  footertext = NULL
)
}
\arguments{
\item{df}{Dataframe to export as table in RMD.}

\item{caption}{Caption for the table.}

\item{fontsize}{Integer value for the fontsize.}

\item{label}{Character string to be used for cross-referencing.}

\item{ignoreCaption}{Caption can be ignored if desired (no caption will be shown for table)}

\item{valueTable}{Defines is table can be used as a "Value Table", allowing to reference
values in the table. A value table requires the presence of a "value" and a "label" column.
If these are not present, an error will occur.}

\item{footertext}{Defined text for the table footer.}
}
\value{
Inline table command as IQRrmd object
}
\description{
Allows to generate RMD code to include a table defined by a dataframe.
To include a table from a table text file, please use the function
rmdTABLE.
}
