% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTABref}
\alias{rmdTABref}
\title{Table Cross reference}
\usage{
rmdTABref(label, caption = NULL)
}
\arguments{
\item{label}{Character string, specifying the label of the table. Labels are limited to
40 characters and should not include whitespaces.}

\item{caption}{A character string with alernative text to show (instead of "Text x").}
}
\value{
Table cross reference command as IQRrmd object
}
\description{
Allows to reference an included table by label.
"@TAB\href{label}{caption}"
}
\details{
If caption ignored the cross-reference will not work.
}
