% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTEXTSTYLE}
\alias{rmdTEXTSTYLE}
\title{Change the size, color, and/or style of a text piece}
\usage{
rmdTEXTSTYLE(text, color = NULL, size = NULL, style = NULL)
}
\arguments{
\item{text}{The text to which the changes should apply.}

\item{color}{string with color name ("blue", "black", "orange", "pink", etc.) In addition to
color names also "#ffeedd" type of definition can be used.}

\item{size}{Numeric value (integer) defining the font size.}

\item{style}{Word character style to use. Note it has to be a Character style! Otherwise nothing is happening!}
}
\value{
Textelement as IQRrmd object
}
\description{
Change the size, color, and/or style of a text piece
}
