% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTITLE}
\alias{rmdTITLE}
\title{Generate title RMD command}
\usage{
rmdTITLE(
  template = "DefaultStyle.rmdt",
  title = "Default Title",
  subtitle = "Default Subtitle",
  date = format(Sys.Date(), "\%d-\%b-\%Y")
)
}
\arguments{
\item{template}{Name of the *.rmdt template file to be used}

\item{title}{Character string with title of the document}

\item{subtitle}{Character string with subtitle of the document}

\item{date}{Character string with date to be shown}
}
\value{
Title IQRrmd object
}
\description{
In the title section of an RMD document the template to be used and the
title page information to be shown are defined. IQnca works with IQReport and
the DefaultStyle.rmdt template document. If you want to use a different template
you can do that but you would need to have one. Contact info@intiquan.com to
learn more. You basically can get IQnca generate Word reports in the desired
style, required for your organization.
}
