% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTXT}
\alias{rmdTXT}
\title{Generate RMD command to include a text from a file}
\usage{
rmdTXT(file, caption = NULL, fontsize = 10, style = "CODE", label = NULL)
}
\arguments{
\item{file}{Character string, specifying the absolute or relative path to the
text file to include.}

\item{caption}{Caption for the text file. If NULL then no caption is shown.}

\item{fontsize}{Integer value for the fontsize.}

\item{style}{Word DOCX template style to use for the text (default: "CODE")}

\item{label}{Character string to be used for cross-referencing. If label and caption
provided, then the caption will be numbered.}
}
\value{
TXT command as IQRrmd object
}
\description{
Generate RMD command to include a text from a file
}
