% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdTXTref}
\alias{rmdTXTref}
\title{Text Cross reference}
\usage{
rmdTXTref(label, caption = NULL)
}
\arguments{
\item{label}{Character string, specifying the label of the text. Labels are limited to
40 characters and should not include whitespaces.}

\item{caption}{A character string with alernative text to show (instead of "Text x").}
}
\value{
Text cross reference command as IQRrmd object
}
\description{
Allows to reference an included text by label.
"@TXT\href{label}{caption}"
}
\details{
If caption ignored the cross-reference will not work.
}
