% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdURL}
\alias{rmdURL}
\title{Generate RMD command to include a URL}
\usage{
rmdURL(url = "http://www.intiquan.com", caption = NULL)
}
\arguments{
\item{url}{Character string, specifying the URL}

\item{caption}{Text to display as hyperlink}
}
\value{
URL command as IQRrmd object
}
\description{
Generate RMD command to include a URL
}
