% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_classes_IQRrmd.R
\name{rmdVALref}
\alias{rmdVALref}
\title{Value Table cross reference}
\usage{
rmdVALref(labelTable, row = NULL, col = NULL, labelValue = NULL)
}
\arguments{
\item{labelTable}{Character string, specifying the label of the table. Labels are limited to
40 characters and should not include whitespaces.}

\item{row}{Numeric value (integer) defining the row of the element to report.
Header row not counted.}

\item{col}{Numeric value (integer) defining the column of the element to report.}

\item{labelValue}{Character string with the element in the "label" column of a
ValueTable. If used then row and col need to be NULL and vice versa.}
}
\value{
Table element value cross reference command as IQRrmd object
}
\description{
Allows to refence (display) values that are stored in labeled tables.
Two ways are possible:
\itemize{
\item Either by referncing row (not counting header row) and column number:
"@VAL\href{labelTable}{row,col}"
\item Or by referencing the label in a ValueTable:
"@VAL\href{labelTable}{labelValue}"
}
}
\details{
Alternative way in RMD (but not implemented in this function) - for ValueTables:
"@VAL\href{labelTable}{labelCell:labelCol:valueCol}"
}
