% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings_plots.R
\name{scale_color_IQnca}
\alias{scale_color_IQnca}
\title{Color table for IQnca}
\usage{
scale_color_IQnca(...)
}
\arguments{
\item{...}{passed to scale_color_manual(...)}
}
\description{
Color table for IQnca
}
\examples{
\dontrun{
df <- data.frame(x = as.factor(sprintf("\%2i",1:19)), y = 1)
IQRggplot(df, aes(x,y, color = x)) +
  geom_point() + scale_color_IQnca()
}
}
