% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings_plots.R
\name{scale_fill_IQnca}
\alias{scale_fill_IQnca}
\title{Color table for IQnca}
\usage{
scale_fill_IQnca(...)
}
\arguments{
\item{...}{passed to scale_fill_manual(...)}
}
\description{
Color table for IQnca
}
