% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_auxiliaries.R
\name{scale_y_log10_IQnca}
\alias{scale_y_log10_IQnca}
\alias{scale_x_log10_IQnca}
\title{Log scale for x or y axis with proper grid lines}
\usage{
scale_y_log10_IQnca(
  labeltype = c("natural", "comma", "scientific", "standard"),
  n_major = 4,
  n_minor = 0,
  ...
)

scale_x_log10_IQnca(
  labeltype = c("natural", "comma", "scientific", "standard"),
  n_major = 4,
  n_minor = 0,
  ...
)
}
\arguments{
\item{labeltype}{character to determine how labels are printed, either "natural" (default),
"comma" (like natural + comma separation for large numbers),
"scientific" (nicely print 10 to the power of ...), or "standard" (standard R).}

\item{n_major}{the number of major labels aiming at. Major labels will always be
powers of 10. With \code{n_major} it is possible to reduce the number of powers shown.}

\item{n_minor}{the exact number of minor labels, i.e. labels between powers of 10.
The argument is only used if the number of major labels drops below \code{n_major}.}

\item{...}{other arguments going to scale_y_log10. If
\code{breaks}, \code{minor_breaks} and \code{labels} are provided, they overwrite the
generated breaks, minor breaks or labels.}
}
\description{
Use as alternative to scale_y_log10().
}
