% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{setup_IQnca}
\alias{setup_IQnca}
\title{Opening the setup_options_IQnca.R file for editing}
\usage{
setup_IQnca(local = FALSE)
}
\arguments{
\item{local}{Logical. If FALSE then the function will open the global file for
editing. If TRUE then the function will open the local file for editing. If
the local file is not present then it will be created. If the definition of
a local file is disabled in the global settings file, the local=TRUE setting
will lead to an error.}
}
\description{
This function opens the setup_options_IQnca.R for editing IQnca options.
The user can choose to edit the general / global settings file in the packages
installation location or create/edit such a file in the users home folder.
On cluster installations a user might not be able to change the global file and
in such a case the use of a users file might be useful.
Note that the users setup_options_IQnca.R file does not need to include all
information present in the global setup_options_IQnca.R file. The user file
can be used to override one or more of the settings in the global file.
In the setup_options_IQnca.R file located in the package installation folder
it can be decided if such a user defined settings file is allowed or not.
}
\concept{Installation}
