% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{slope_IQdataNCA}
\alias{slope_IQdataNCA}
\title{Slope calculation function for IQdataNCA objects}
\usage{
slope_IQdataNCA(data, manualslope = NULL)
}
\arguments{
\item{data}{IQdataNCA object}

\item{manualslope}{List of entries. One entry per subject/profile combination, allowing to define the
slope points. A list entry in this list can be generated using the function slope_manual_NCA().}
}
\value{
Updated IQdataNCA object
}
\description{
Calculates the slope information based on the data. By default the bestslope
algorithm is used but for indidivual subjects also the slopepoints can be defined that
should be used.
}
\seealso{
Other NCA Slope: 
\code{\link{slope_manual_NCA}()},
\code{\link{slopetest_IQnca}()}
}
\concept{NCA Slope}
