% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{slope_manual_NCA}
\alias{slope_manual_NCA}
\title{Define slope points in subjects to be used for manual slope calculation}
\usage{
slope_manual_NCA(USUBJID, PROFILE = NULL, IXslope)
}
\arguments{
\item{USUBJID}{Single USUBJID}

\item{PROFILE}{Name of the profile (PROFILE column)}

\item{IXslope}{Numeric vector with slope points matching the IX column entries in the IQdataNCA object}
}
\value{
Just a list  - can be used as input argument 'manualslope' to slope_IQdataNCA()
}
\description{
Helper function to construct the manualslope input argument for slope_IQdataNCA()
}
\seealso{
Other NCA Slope: 
\code{\link{slope_IQdataNCA}()},
\code{\link{slopetest_IQnca}()}
}
\concept{NCA Slope}
