% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{slopetest_IQnca}
\alias{slopetest_IQnca}
\title{Calculation of R2 and R2ADJ for slope of profile}
\usage{
slopetest_IQnca(
  data,
  USUBJID,
  PROFILE = NULL,
  IXslope = NULL,
  logY = TRUE,
  R2ADJTHRESHOL = 0.85
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{USUBJID}{Unique subject ID of subject to do the slope calculation for}

\item{PROFILE}{Name of the profile. Only needs to be provided when USUBJID appears in more than
one profile in the data}

\item{IXslope}{Vector with manually selected numeric IX values in the data to use for slope calculation
or set to NULL the bestSlope algorithm will be used}

\item{logY}{Log y axis is TRUE, otherwise linear Y axis}

\item{R2ADJTHRESHOL}{Threshold for R2ADJ to switch color and annotation}
}
\description{
This function calculates the R2 (R-squared) and R2ADJ (R2-squared adjusted) metrics
for the slope fitting. User provided indices for PK samples to be considered will be used
if provided. If not provided, the bestSlope algorithm will be used. The result is presented
in a graphical manner.
}
\details{
The bestslope algorithm considers points from CMAX/TMAX until the last >0 point.
For extravascular one point later is used as start point. For INFUSION when >= 5
samples available also one point later is started.
The algorithm simply determines metrics from all points to the last point with minimally
3 points in the calculation. The resulting best slope is selected as the solution with
the maximum number of used concentration points for slope calculation for which the
R2ADJ is not smaller than SLOPETOL (in the data) as compared to the maximum R2ADJ.
}
\seealso{
Other NCA Slope: 
\code{\link{slope_IQdataNCA}()},
\code{\link{slope_manual_NCA}()}
}
\concept{NCA Slope}
