% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{statXYnominal}
\alias{statXYnominal}
\title{Summary stats per time point (no clustering)}
\usage{
statXYnominal(x, y)
}
\arguments{
\item{x}{x/time vector}

\item{y}{value vector}
}
\value{
statistics (mean, median, etc. for each time point).
}
\description{
Summary stats per time point (no clustering)
}
