% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{stopIQR}
\alias{stopIQR}
\title{Replacement for stop function to be used in IQnca}
\usage{
stopIQR(..., call. = FALSE, domain = NULL)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character (and which are pasted together with no separator) or a single condition object.}

\item{call.}{logical, indicating if the call should become part of the error message.}

\item{domain}{see documentation or same arument in stop() function}
}
\description{
Replacement for stop function to be used in IQnca
}
