% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{summary.IQdataNCA}
\alias{summary.IQdataNCA}
\title{Generic summary function for IQdataNCA objects}
\usage{
\method{summary}{IQdataNCA}(object, ...)
}
\arguments{
\item{object}{IQdataNCA object}

\item{...}{Additional arguments - whatever R hides in these ... bad style!}
}
\description{
Generic summary function for IQdataNCA objects
}
