% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{summary.IQRoutputFigure}
\alias{summary.IQRoutputFigure}
\title{Summary function for IQRoutputFigure}
\usage{
\method{summary}{IQRoutputFigure}(object, ...)
}
\arguments{
\item{object}{IQRoutputFigure object}

\item{...}{Additional arguments (unused)}
}
\description{
Prints information on IQRoutputFigure with no return.
}
\author{
Anne Kümmel, IntiQuan
}
