% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings_additional_classes_IQdataNCA.R
\name{summaryObservations_IQdataNCA}
\alias{summaryObservations_IQdataNCA}
\title{Table summarizing the available observations}
\usage{
summaryObservations_IQdataNCA(
  data,
  stratifyBy = "PROFILE",
  labels = NULL,
  tableTitle = NULL,
  footerAddText = NULL,
  filename = NULL,
  FLAGpatients = FALSE,
  FLAGtotal = TRUE,
  digits = 4,
  report = NULL
)
}
\arguments{
\item{data}{IQncaData}

\item{stratifyBy}{character vector defining the columns used for stratification (default: \code{Profile})}

\item{labels}{Named vector with labels that should be used as names for stratification column(s)}

\item{tableTitle}{Table title}

\item{footerAddText}{Note to be added to table footer}

\item{filename}{filename to print table to}

\item{FLAGpatients}{whether to use the term "patients" instead of "subjects"}

\item{FLAGtotal}{whether a row with the total numbers should be added}

\item{digits}{Number of significant digits for rounding of percentages}

\item{report}{Argiment to IQRoutputTable whether to prepare the table for reporting}
}
\value{
IQRoutputTable if no filename given. If filename is given, table is printed to file and nothing is returned.
}
\description{
Statistics on number of subjects, samples, records below LLOQ, and ignored records are
calculated and presented. Counts are stratified per default by profile.
}
\examples{
\dontrun{
summaryObservations_IQdataNCA(
  data = dataNCA,
  stratifyBy    = c("GROUP", "ANALYTE"),
  labels = c("GROUP" = "Dose", "ANALYTE" = "Analyte"),
  filename = "summary_observations.txt"
)
}
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()},
\code{\link{reportIgnoredSUMNCA_IQdataNCA}()}
}
\concept{NCA Data}
