% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{support}
\alias{support}
\title{Open support webpage}
\usage{
support()
}
\description{
Convenient function to open the support webpage
}
\seealso{
Other Help & Documentation: 
\code{\link{doc_IQnca}()},
\code{\link{examples_IQnca}()},
\code{\link{install_MIDDmodule}()},
\code{\link{install_workshop}()}
}
\concept{Help & Documentation}
