% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{table_summary_acrossGROUP_pkparameters_IQnca}
\alias{table_summary_acrossGROUP_pkparameters_IQnca}
\title{Generate a summary table for PK parameters (across dose groups)}
\usage{
table_summary_acrossGROUP_pkparameters_IQnca(
  data,
  tablenumber = NULL,
  strat = NULL,
  fontsizetable = 8,
  filename = "table_acrossgroup_PKparam",
  SIGNIF = 4,
  parameterReport = "standard"
)
}
\arguments{
\item{data}{IQnca object}

\item{tablenumber}{Character string with table number information, added to the title for each table, if defined}

\item{strat}{Name of a column in the NCA result dataset (data) to use for stratification}

\item{fontsizetable}{Fontsize to be used in the table}

\item{filename}{Filename to export the listing IQRrmd object to}

\item{SIGNIF}{Significant digits for values}

\item{parameterReport}{Character string defining the PK parameters to report in the table
By default these parameters will be selected based on the "Standard" parameters defined in the IQparamNCA_Specification.xlsx
document, subject to the profile and administration type. The user can select "All" or "Standard".
Alternatively, parameterReport can also be a vector or parameter names (based on PKPARAMCD in IQparamNCA_Specification.xlsx).
Instead of the standard or all ones then these will be considered in the listing (also subject to profile and
administration type). Note that interval AUC (if calculated is always added by default!)}
}
\value{
An IQRrmd object exported to a file
}
\description{
The table will be generated already as an IQRrmd object, allowing easy reporting in Word with IQReport.
Stratification by PROFILE and a potential additional stratification column. Grouping across GROUP (dose)
will always be done. Meaning that only dose independent parameters should be considered.
}
\seealso{
Other NCA Parameter Tables: 
\code{\link{reliabilityTable_IQnca}()},
\code{\link{table_summary_byGROUP_pkparameters_IQnca}()}
}
\concept{NCA Parameter Tables}
