% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_classes_IQdataNCA.R
\name{table_summary_PKconc_IQdataNCA}
\alias{table_summary_PKconc_IQdataNCA}
\title{Generate a table of statistical summaries of pharmacokinetic concentrations}
\usage{
table_summary_PKconc_IQdataNCA(
  data,
  stratify_by = c("GROUPN", "PROFILE"),
  statistics = c("MEAN", "SD", "CVPMN", "GMEAN", "CVPGM", "MEDIAN", "MIN", "MAX"),
  table_number = NULL,
  table_split_by = NULL,
  table_head = c("AGE", "RACE", "SEX"),
  table_cols = c("PROFILE", "GROUP"),
  table_compare = NULL,
  labels = NULL,
  roundfun = signif,
  digits = c(MEAN = 3, SD = 3, CVPMN = 3),
  fontsize = 8,
  page_orientation = c("landscape", "portrait"),
  filename = NULL
)
}
\arguments{
\item{data}{IQdataNCA object}

\item{stratify_by}{Columns that define the groups that are summarized}

\item{statistics}{Statistics to be calculated and displayed. Defined by charater vector of shortnames:
\tabular{ll}{
Short name   \tab Description  \cr
============ \tab ===========  \cr
N            \tab Number of subjects \cr
MEAN         \tab Arithmetic mean \cr
SD           \tab Arithmetic standard deviaion \cr
CVPMN        \tab Arithmetic coefficient of variation (percent) \cr
CI95MN       \tab 95\% confidence interval for arithmetic mean \cr
CI90MN       \tab 90\% confidence interval for arithmetic mean \cr
GMEAN        \tab Geometric mean \cr
GSD          \tab Geometric standard deviaion \cr
CVPGM        \tab Geometric coefficient of variation (percent) \cr
CI95GM       \tab 95\% confidence interval for geometric mean \cr
CI90GM       \tab 90\% confidence interval for geometric mean \cr
MEDIAN       \tab Median \cr
MIN          \tab Maximum value \cr
MAX          \tab Minimum value \cr
}}

\item{table_number}{Character string with table number information, added to the title for each table, if defined}

\item{table_split_by}{Column to split tables by (defaults to NULL, hence, only single table produced)}

\item{table_head}{Common information for split to be displayed above table}

\item{table_cols}{Columns to be displayed before the summary statistics}

\item{table_compare}{Single column defining groups for which the statistics are displayed side-by-side}

\item{labels}{Customized labels for columns that substitute the column name in the IQdataNCA object. Applies to 'table_head', 'table_cols', and the entries of 'table_compare'}

\item{roundfun}{Rounding function to be used (defaults to \link{signif})}

\item{digits}{Number of digits to be used for rounding}

\item{fontsize}{Fontsize to be used in the table}

\item{page_orientation}{Page orientation, 'landscape' or 'portrait'}

\item{filename}{Filename to export the table IQRrmd object to}
}
\value{
An IQRrmd object exported to a file
}
\description{
The table will be generated already as an IQRrmd object, allowing easy reporting in Word with IQReport.
The table is done per PROFILE and GROUP.
Important: Ignored records (INGORER & IGNORSUM) are not reported in these tables! Ignored subjects (IGNOREI) are
removed as well.
}
\examples{
\dontrun{
pending
}
}
\seealso{
Other TLF customized: 
\code{\link{figure_indivConc_IQdataNCA}()},
\code{\link{figure_lines_IQdataNCA}()},
\code{\link{figure_summary_IQdataNCA}()},
\code{\link{listing_PKconc_IQdataNCA}()},
\code{\link{listing_PKpars_IQnca}()},
\code{\link{table_summary_PKpars_IQnca}()}
}
\concept{TLF customized}
