% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_classes_IQdataNCA.R
\name{table_summary_PKpars_IQnca}
\alias{table_summary_PKpars_IQnca}
\title{Generate a table of statistical summaries of pharmacokinetic parameters}
\usage{
table_summary_PKpars_IQnca(
  data,
  stratify_by = c("GROUPN", "PROFILE"),
  statistics = c("MEAN", "SD", "CVPMN", "GMEAN", "CVPGM", "MEDIAN", "MIN", "MAX"),
  table_number = NULL,
  table_split_by = NULL,
  table_head = c("COMPOUND", "MATRIX", "ANALYTE"),
  table_cols = c("PROFILE", "GROUP"),
  table_pkpars = "standard",
  labels = NULL,
  roundfun = signif,
  digits = c(MEAN = 3, SD = 3, CV = 3),
  fontsize = 8,
  page_orientation = c("landscape", "portrait"),
  filename = NULL
)
}
\arguments{
\item{data}{IQnca object with individual PK parameters}

\item{stratify_by}{Columns that define the groups that are summarized}

\item{statistics}{Statistics to be calculated and displayed. Defined by charater vector of shortnames:
\tabular{ll}{
Short name   \tab Description  \cr
============ \tab ===========  \cr
N            \tab Number of subjects \cr
MEAN         \tab Arithmetic mean \cr
SD           \tab Arithmetic standard deviaion \cr
CVPMN        \tab Arithmetic coefficient of variation (percent) \cr
CI95MN       \tab 95\% confidence interval for arithmetic mean \cr
CI90MN       \tab 90\% confidence interval for arithmetic mean \cr
GMEAN        \tab Geometric mean \cr
GSD          \tab Geometric standard deviaion \cr
CVPGM        \tab Geometric coefficient of variation (percent) \cr
CI95GM       \tab 95\% confidence interval for geometric mean \cr
CI90GM       \tab 90\% confidence interval for geometric mean \cr
MEDIAN       \tab Median \cr
MIN          \tab Maximum value \cr
MAX          \tab Minimum value \cr
}}

\item{table_number}{Character string with table number information, added to the title for each table, if defined}

\item{table_split_by}{Column to split tables by (defaults to NULL, hence, only single table produced)}

\item{table_head}{Common information for split to be displayed above table}

\item{table_cols}{Columns to be displayed before the summary statistics}

\item{table_pkpars}{Character string defining the PK parameters to report in the table
By default these parameters will be selected based on the "Standard" parameters defined in the IQparamNCA_Specification.xlsx
document, subject to the profile and administration type. The user can select "All" or "Standard".
Alternatively, parameterReport can also be a vector or parameter names (based on PKPARAMCD in IQparamNCA_Specification.xlsx).
Instead of the standard or all ones then these will be considered in the listing (also subject to profile and
administration type).}

\item{labels}{Customized labels for columns that substitute the column name in the IQdataNCA object. Applies to 'table_head', 'table_cols', and the entries of 'table_compare'}

\item{roundfun}{Rounding function to be used (defaults to \link{signif})}

\item{digits}{Number of digits to be used for rounding}

\item{fontsize}{Fontsize to be used in the table}

\item{page_orientation}{Page orientation, 'landscape' or 'portrait'}

\item{filename}{Filename to export the table IQRrmd object to}
}
\value{
An IQRrmd object exported to a file or IQncaTableList
}
\description{
PK parameters are summarized by a set of statistical values.
The function acknowleges the reliability of NCA parameters
and only includes reliable ones.
}
\examples{
\dontrun{
pending
}
}
\seealso{
Other TLF customized: 
\code{\link{figure_indivConc_IQdataNCA}()},
\code{\link{figure_lines_IQdataNCA}()},
\code{\link{figure_summary_IQdataNCA}()},
\code{\link{listing_PKconc_IQdataNCA}()},
\code{\link{listing_PKpars_IQnca}()},
\code{\link{table_summary_PKconc_IQdataNCA}()}
}
\concept{TLF customized}
