% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_tools.R
\name{text_IQRoutputTable}
\alias{text_IQRoutputTable}
\title{Convert IQRoutputTable to text representation}
\usage{
text_IQRoutputTable(
  object,
  report = NULL,
  na.string = NULL,
  FLAGreplaceRoundBracketsHeader = TRUE
)
}
\arguments{
\item{object}{IQRoutputTable object}

\item{report}{TRUE: add reporting tags
FALSE: do not add (for scfreen output only).
Also stored in object but can be overridden here}

\item{na.string}{Character string to replace NA entries by. Defaults to NULL.}

\item{FLAGreplaceRoundBracketsHeader}{If TRUE then "(" and ")" in table headers will be replaced
by \code{"["} and \code{"["}. If FALSE, then no such replacement is done. Needed for compatibility with IQReport on Linux.}
}
\value{
Character string with table in text format
}
\description{
Convert IQRoutputTable to text representation
}
