% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_TLFs.R
\name{textRMD_IQncaTableList}
\alias{textRMD_IQncaTableList}
\title{Function to translate listings or tables (IQncaTableList object) to rmd}
\usage{
textRMD_IQncaTableList(
  tabs,
  fontsize = 8,
  page_orientation = c("landscape", "portrait")
)
}
\arguments{
\item{tabs}{IQncaTableList object}

\item{fontsize}{Fontsize to be used for table font}

\item{page_orientation}{Page orientation, 'landscape' or 'portrait'}
}
\value{
r
}
\description{
Function to translate listings or tables (IQncaTableList object) to rmd
}
