% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_settings_plots.R
\name{themeIQnca}
\alias{themeIQnca}
\title{IQ NCA's standard theme}
\usage{
themeIQnca(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{numeric, font-size}

\item{base_family}{character, font-name}
}
\description{
The standard gg theme is based on \code{theme_bw}.
}
