% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprogramming_tools.R
\name{unlabel_dataframe}
\alias{unlabel_dataframe}
\title{Remove labelled class and label}
\usage{
unlabel_dataframe(data, removeLabelledClass = TRUE, removeLabel = FALSE)
}
\arguments{
\item{data}{data.frame with potentially labeled columns}

\item{removeLabelledClass}{TRUE removes the class "labelled"}

\item{removeLabel}{TRUE removes the label attribute}
}
\value{
data.frame with Unlabelled cols
}
\description{
Remove labelled class and label
}
