% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_auxiliaries.R
\name{warningIQR}
\alias{warningIQR}
\title{Replacement for warning function to be used in IQnca}
\usage{
warningIQR(..., call. = FALSE, domain = NULL)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character (and which are pasted together with no separator) or a single condition object.}

\item{call.}{logical, indicating if the call should become part of the warning message.}

\item{domain}{see documentation or same arument in warning() function}
}
\description{
Replacement for warning function to be used in IQnca
}
