% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_classes_IQdataNCA.R
\name{write_IQncaFigure}
\alias{write_IQncaFigure}
\title{Function to plot IQncaFigure}
\usage{
write_IQncaFigure(figs, filename, ...)
}
\arguments{
\item{figs}{IQncaFigure object}

\item{filename}{File to print the figures to.}

\item{...}{Arguments to \link{IQRoutputFigure} to adjust plotting settings}
}
\value{
Nothing. Output to files.
}
\description{
A PDF with the of the figure(s) is produced for the given filename. This will contain the title and footers.
Next to the file, a folder with meta information (e.g., the rmd file and adjusted pdf) is produced. The rmd and the
adjusted PDF can be used to generate a report via IQReport.
}
