% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_classes_TLFs.R
\name{write_IQncaTableList}
\alias{write_IQncaTableList}
\title{Writes rmd file for reporting of listing or table in a IQReport-compiled Word document}
\usage{
write_IQncaTableList(tabs, filename, fontsize = 8, page_orientation)
}
\arguments{
\item{tabs}{IQncaTableList object containing a list of tables (IQncaTable objects) with titles and footers}

\item{filename}{Filename towrite the rmd to}

\item{fontsize}{Font to be used in tables}

\item{page_orientation}{whether to plot table on landscape or portrait page}
}
\value{
Nothing return. Output to file.
}
\description{
Writes rmd file for reporting of listing or table in a IQReport-compiled Word document
}
\examples{
\dontrun{
write_IQncaTableList(
 tableList,
 filename = "../Output/tables/mytable.rmd",
 page_orientation = "landscape"
)
}
}
