% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_classes_IQRdataNCA.R
\name{reportIgnoredSUMNCA_IQdataNCA}
\alias{reportIgnoredSUMNCA_IQdataNCA}
\title{Report ignored records from summary tables and/or NCA parameter determination}
\usage{
reportIgnoredSUMNCA_IQdataNCA(data, filename = NULL)
}
\arguments{
\item{data}{IQdataNCA object for which to do the reporting}

\item{filename}{Filename, possibly including path, to store are IQRoutputTable text file for Word reporting}
}
\description{
Reports records that are ignored from either summary tables or NCA parameter determination
with the reason why. Displays on screen and can also save the information to a file as an IQRoutputTable object.
}
\seealso{
Other NCA Data: 
\code{\link{IQdataNCA}()},
\code{\link{check_IQdataNCA}()},
\code{\link{dataSpec_IQdataNCA}()},
\code{\link{exportADNCA_IQdataNCA}()},
\code{\link{export_IQdataNCA}()},
\code{\link{ignoreNCA_IQdataNCA}()},
\code{\link{ignoreSUM_IQdataNCA}()},
\code{\link{ignorerecord_IQdataNCA}()},
\code{\link{ignoresubject_IQdataNCA}()},
\code{\link{load_IQdataNCA}()},
\code{\link{plot.IQdataNCA}()}
}
\concept{NCA Data}
